Split test/process.js into test/process/ per-process files
CodeScene flagged test/process.js (9.09) for having 425 functions in a
single module (a mild God File warning). Split it into a test/process/
directory, one file per process, mirroring the existing test/mc/
per-sampler layout:

- process.js: the abstract Process base class tests
- brownian-motion.js, geometric-brownian-motion.js,
  ornstein-uhlenbeck.js, brownian-bridge.js, ar1.js,
  cox-ingersoll-ross.js, random-walk.js: one file per process
- poisson.js: Poisson plus its deprecated PoissonProcess alias
- compound-poisson.js: CompoundPoisson plus its deprecated
  CompoundPoissonProcess alias
- reference-values.js: the process-cases.js-driven analytical
  reference-value suite
- _helpers.js: shared sampling/assertion helpers (assertSampleMoments,
  MOMENT_SEEDS, K_SIGMA, etc.)

Every new file scores a Code Health 10.0. Updated the stale rationale
comment in test/process-cases.js and CLAUDE.md's Testing Conventions
section to describe the new layout. No behavior change: same test
cases, same assertions, all 10146 tests still pass.
