+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `docs/accuracy.md`: a committed, documented-accuracy-bounds table generated by the new `scripts/generate-accuracy-docs.js` (`npm run accuracy`, chaining `accuracy:special` → `accuracy:dist` → `accuracy:docs`) from the `#1264`/`#1265` differential-testing harness JSON reports, closing the "Documented accuracy bounds" gap `todo.md` tracked under Publication-Grade Gaps. Every special function `src/special/index.js` exports and every distribution `src/dist/index.js` exports is listed — swept ones with their measured domain (read straight from the harness report's own `domain` field, so the table can never drift from what was actually sampled), max/median ULP, and sample count; unswept ones as an explicit "not yet measured" row rather than a silent omission. `Gamma.pdf`/`InverseGamma.pdf` (#1265's NaN/overflow defects, now #1363/#1364) and `besselK`/`besselKnu`'s `x=6` series/asymptotic crossover (#1140) render with their actual measured (bad) values — `∞` for the NaN-mismatch divergences, ULP counts in the billions for the Bessel crossover — each linked to its tracking issue, never rounded away. `docs/accuracy.md` is committed rather than generated at docs-build time, so it is readable on GitHub without a Python + mpmath environment and its diffs are reviewable per-PR; see [ADR-0053](decisions/0053-accuracy-docs-committed.md). Both harness scripts' `build_report()` gained a `domain`/`mp_dps` field (read from the same `SWEEP_SPEC`/`DIST_SPEC` dict the sweep itself draws from) to support this without risking the "for `|x| <= Y`" claim drifting from what was actually measured. `generate-accuracy-docs.js`'s `statusFor()` now composes every applicable flag instead of stopping at the first match: a divergence count no longer disappears once an entry also carries a `KNOWN_ISSUES` link (`Gamma.pdf`'s row now reads "31 divergence(s) ... known accuracy gap" instead of swallowing the count), and a report's `errors` field (points where the harness's Node eval bridge threw rather than returning a value — a distinct, more severe failure mode than a returned-but-wrong value) is now rendered too, previously tracked in the JSON report but never surfaced in the table at all. Each of the four flags (thrown error, divergence, ceiling breach, known-issue link) now gets an emoji chosen for what it means rather than a shared `⚠️` or an arbitrary severity color — 💥 the eval crashed, ❌ a value came back but is NaN/nonsensical against mpmath, ⚠️ a real finite value worse than its calibrated ceiling, 🔗 a pointer to an already-tracked non-new problem, ✅ OK — so the failure mode reads at a glance without parsing the status text.
 - `.github/workflows/difftest.yml`: runs the `#1264`/`#1265` differential-testing harness on a weekly schedule (plus `workflow_dispatch` for manual runs), separately from `ci.yml` — the harness needs a Python + mpmath environment and sweeps far denser grids than `test/precision-*.js`, so it stays out-of-band from the fast, merge-blocking unit-test gate ([ADR-0052](decisions/0052-differential-testing-harness-live-mpmath-out-of-band.md)). The job runs `accuracy:special`/`accuracy:dist`, uploads both JSON reports as a workflow artifact, then runs the new `scripts/difftest-ci-gate.js` (`npm run difftest:ci-gate`), which fails the job when any function/distribution exceeds its declared `ulp_ceiling` and writes the exact reproducer — function/distribution, parameter tuple, evaluation point, ranjs value, mpmath value, ULP distance — to the job summary. Deliberately does not auto-file or update a tracking issue on failure: a red run plus the uploaded report is a sufficient signal, and a per-run auto-filed issue would duplicate weekly on top of what `stale.yml` already manages. The gate is a separate script rather than added to the harness scripts themselves, since `npm run difftest:*` also doubles as a plain local diagnostic run that should not start failing the shell over an already-tracked, already-calibrated defect (e.g. `besselK`'s #1140 crossover). `difftest-ci-gate.js` now also fails the job on `divergences > 0` or `errors > 0`, not only `ceiling_exceeded` — an `inf` ULP distance (the harness's encoding for a NaN/Infinity mismatch against a finite mpmath reference) is deliberately excluded from `ceiling_exceeded`'s comparison, so without this a regression to NaN/Infinity on a previously-clean function/distribution would have stayed invisible to the gate (#1369). A `KNOWN_ISSUES` allowlist (mirroring `generate-accuracy-docs.js`'s own map) keeps the two already-tracked divergence sources, `Gamma.pdf` (#1363) and `InverseGamma.pdf` (#1364), from turning the job permanently red on ship; allowlisted entries still appear in the job summary table, tagged with their tracking issue, rather than silently disappearing. The summary table gained a `Reason` column so a divergence or eval-error failure reads distinctly from a plain ceiling breach instead of leaving a reviewer to guess why a row with a small `max_ulp` still failed.
 
+### Changed
+
+- Code Health of `test/process.js` improved from 9.09 to 10.0 by splitting it into a `test/process/` directory, one file per process (mirroring the existing `test/mc/` per-sampler layout), plus a shared `_helpers.js` and a `reference-values.js` consuming `test/process-cases.js`. No behavior change — same test cases, same assertions.
+
 ### Fixed
 
 - `ran.special.besselKnu(nu, x)`: silently returned up to ~77% relative error for orders `nu`
