+++ CLAUDE.md
@@ -142,9 +142,9 @@ node scripts/check-subpath-exports.js  # package.json subpath in sync
 
 | Item | File | What to add |
 | --- | --- | --- |
-| Reference values | `test/process-cases.js` | Entry with `name`, `ctor`, and `refs`. Each ref carries `should` (prose completing the mocha title), `params` (a **thunk** returning constructor arguments, so no instance — and no PRNG — is built at import), `method`, `args`, optional `chain`/`chainArgs` (for `marginal(t).pdf(x)`), `expected`, `tol`, and `source`. `source` is passed as the assertion message, so provenance shows up in failure output instead of only in a comment a reader has to find. Consumed by a plain inline `forEach` at the end of `test/process.js` — there is no `dist-runner.js`-style shared runner, and at 9 processes there should not be one. |
+| Reference values | `test/process-cases.js` | Entry with `name`, `ctor`, and `refs`. Each ref carries `should` (prose completing the mocha title), `params` (a **thunk** returning constructor arguments, so no instance — and no PRNG — is built at import), `method`, `args`, optional `chain`/`chainArgs` (for `marginal(t).pdf(x)`), `expected`, `tol`, and `source`. `source` is passed as the assertion message, so provenance shows up in failure output instead of only in a comment a reader has to find. Consumed by a plain inline `forEach` in `test/process/reference-values.js` — there is no `dist-runner.js`-style shared runner, and at 9 processes there should not be one. |
 | Precision gate | `test/precision-process.js` | Generated — see below. |
-| Behavioural tests | `test/process.js` | Contract behaviour the precision gate does not probe: `NaN` for `t <= 0`, exact `0` outside the support, `Infinity` at a collapsed endpoint, symmetry about the mean. Sampled moments go through `assertSampleMoments` at the fixed `MOMENT_SEEDS = [0, 42, 12345]`, compared against a closed form derived independently from the SDE with a CLT-derived `K_SIGMA = 8` tolerance — never against the process's own `mean()`/`variance()`, which would be a tautology. |
+| Behavioural tests | `test/process/<name>.js` | Contract behaviour the precision gate does not probe: `NaN` for `t <= 0`, exact `0` outside the support, `Infinity` at a collapsed endpoint, symmetry about the mean. Sampled moments go through `assertSampleMoments` at the fixed `MOMENT_SEEDS = [0, 42, 12345]`, compared against a closed form derived independently from the SDE with a CLT-derived `K_SIGMA = 8` tolerance — never against the process's own `mean()`/`variance()`, which would be a tautology. Both helpers, along with the other shared test utilities in that describe block, live in `test/process/_helpers.js`. |
 | Subpath export | `package.json` `exports` field | `"./process/<name>": { "import": "./dist/process/<name>.esm.js" }` in alphabetical order. Verified by `node scripts/check-subpath-exports.js`, which covers `process` and `mc` as well as `dist`. |
 | Named export | `src/process/index.js` | Add the export line |
 | CHANGELOG entry | `CHANGELOG.md` `## [Unreleased]` section | Mandatory — a new process is user-visible |
@@ -193,7 +193,8 @@ Every public function and method signals "no ordinary result" through exactly on
 - `test/dist-runner.js` — shared `registerDistributionTests(testCases)` used by the shard files to register the constructor/seed/load-save/analytical/sample/test/moments/fit tree for each case.
 - `test/dist-shard-0.js` .. `test/dist-shard-3.js` — deterministically partition `[...continuousCases, ...discreteCases]` by index modulo 4 and run their slice through `dist-runner.js`, so `mocha --parallel` can execute the full distribution suite across worker processes.
 - `test/dist-base-*.js` — the non-distribution-suite tests, split out of a former monolithic `dist-base.js` to stay under CodeScene's per-file Code Health thresholds (see issue #1072): `dist-base-core.js` (base `Distribution` instance methods other than `.fit()`/`.params()`), `dist-base-fit-1.js`/`-fit-2.js`/`-fit-3.js` (the `.fit()`/`_fitInit` block, split into three roughly-equal parts), `dist-base-params.js` (`.params()`), `dist-base-special-cases.js` (`PreComputed`, `Degenerate`, `Kolmogorov`, `_qEstimateWalk`), `dist-base-paramcount.js` (Lindley/Mielke/Davis and the multi-level-inheritance parameter-count regressions), `dist-base-fit-gate.js` (the `fit()` precision and robustness gate), and `dist-base-sharding.js` (the `dist-shard-*.js` partition regression guard).
-- `test/process-cases.js` — externally-sourced reference values for `ran.process` analytical methods (issue #1221), consumed by a plain inline `forEach` at the end of `test/process.js`.
+- `test/process/` — one file per process (`brownian-motion.js`, `ar1.js`, etc.), a `process.js` for the abstract `Process` base class, a `_helpers.js` for shared sampling/assertion helpers (`assertSampleMoments`, `MOMENT_SEEDS`, `K_SIGMA`, etc.), and `reference-values.js` consuming `test/process-cases.js` — mirrors the `test/mc/` per-sampler layout.
+- `test/process-cases.js` — externally-sourced reference values for `ran.process` analytical methods (issue #1221), consumed by a plain inline `forEach` in `test/process/reference-values.js`.
 - `test/precision-continuous.js`, `test/precision-discrete.js`, `test/precision-special.js`, `test/precision-summary-stats.js`, `test/precision-process.js` — precision gates pairing a `REFS` array of mpmath-derived reference values with an inline `forEach` consumer in the same file, rather than a separate runner module.
 - **Shared runner vs. data-file-plus-inline-loop**: the codebase has two structured-case patterns. Use a shared runner module (`dist-runner.js`) only when more than one file consumes the same cases, as the four `dist-shard-*.js` files do. Otherwise pair a plain array of case objects with an inline `forEach` that builds `describe`/`it` directly, as `test/process-cases.js` and the five `test/precision-*.js` files do — a runner with a single caller is an abstraction boundary nothing needs. See the rationale recorded in the header comment of `test/process-cases.js`.
 - **Behavior-first assertions**: assert on the output of public methods given known inputs (hand-calculated expected values), not on internal state.
