--- test/process.js
@@ -1,3151 +1,0 @@
-import { assert } from 'chai'
-import { describe, it } from 'mocha'
-import Process from '../src/process/_process'
-import AR1 from '../src/process/ar1'
-import BrownianBridge from '../src/process/brownian-bridge'
-import BrownianMotion from '../src/process/brownian-motion'
-import CoxIngersollRoss from '../src/process/cox-ingersoll-ross'
-import GeometricBrownianMotion from '../src/process/geometric-brownian-motion'
-import OrnsteinUhlenbeck from '../src/process/ornstein-uhlenbeck'
-import CompoundPoisson from '../src/process/compound-poisson'
-import CompoundPoissonProcess from '../src/process/compound-poisson-process'
-// Aliased to avoid colliding with ran.dist.Poisson (imported below and used as a jump
-// distribution in the CompoundPoisson tests) — the two are unrelated classes that happen
-// to share a bare name (decisions/0041-process-subclass-naming-no-process-suffix.md).
-import ProcessPoisson from '../src/process/poisson'
-import PoissonProcess from '../src/process/poisson-process'
-import RandomWalk from '../src/process/random-walk'
-import ShiftedBinomial from '../src/dist/_shifted-binomial'
-import { Normal, LogNormal, Gamma, Poisson, Tweedie } from '../src/dist'
-import { chiTest } from './test-utils'
-import processCases from './process-cases'
-
-// Fixed seeds replace ksTest/chiTest significance-level checks: a random seed can produce
-// a false positive/negative at the chosen critical value on some runs, while a fixed seed
-// deterministically reproduces the same sample every run.
-const MOMENT_SEEDS = [0, 42, 12345]
-
-// The scipy/`Python3 math` reference values asserted at 1e-10 in the .pdf() and .marginal()
-// blocks below are deliberately KEPT rather than superseded by test/precision-process.js
-// (issue #1223), which now gates the same densities at 1e-14 against mpmath mp.dps=50
-// references over a systematic 3-parameter-sets x 3-times x 5-interior-points grid.
-// They are not redundant: each block also pins contract behavior the precision gate does
-// not probe — NaN for t <= 0, exact 0 outside the support, Infinity at the Brownian
-// bridge's collapsed endpoints, and symmetry about the mean — and the inline scipy
-// derivations document the closed form at the point of use. Treat precision-process.js,
-// not these, as the authority on how many digits a process density is required to match.
-
-// Sample mean/variance are compared against an exact closed-form derived independently from
-// each process's SDE/update rule — never against the process's own mean()/variance() methods,
-// which would make the test a tautology against the code under test (decisions: never write
-// the same formula in both the production method and the test assertion). The comparison uses
-// a CLT-derived tolerance: SE(mean) = sqrt(variance/n), and SE(variance) ≈ variance*sqrt(2/(n-1))
-// (exact for Gaussian data, a reasonable order-of-magnitude bound for the Poisson/Bernoulli/
-// Gamma-like samples used elsewhere in this file). K standard errors keeps false failures
-// negligible at all three fixed seeds.
-const K_SIGMA = 8
-
-function assertSampleMoments (sample, expectedMean, expectedVariance, seed) {
-  const n = sample.length
-  const mean = sample.reduce((a, b) => a + b, 0) / n
-  const variance = sample.reduce((a, b) => a + (b - mean) * (b - mean), 0) / (n - 1)
-  const tolMean = K_SIGMA * Math.sqrt(expectedVariance / n)
-  const tolVariance = K_SIGMA * expectedVariance * Math.sqrt(2 / (n - 1))
-  assert.closeTo(mean, expectedMean, tolMean, `seed ${seed}: sample mean ${mean} vs expected ${expectedMean}`)
-  assert.closeTo(variance, expectedVariance, tolVariance, `seed ${seed}: sample variance ${variance} vs expected ${expectedVariance}`)
-}
-
-// Draws n single-step samples from a continuously evolving path, transforming each
-// (newState, prevState) pair. Relies on the step distribution being stationary and
-// independent of the current state (true for all processes this helper is used with).
-function sampleSteps (proc, n, transform = (curr, prev) => curr - prev) {
-  const samples = []
-  for (let i = 0; i < n; i++) {
-    const prev = proc.state()
-    proc.next()
-    samples.push(transform(proc.state(), prev))
-  }
-  return samples
-}
-
-// Draws n independent single-step samples from x0, resetting between draws. Needed
-// whenever the step distribution depends on the current state (e.g. a time-indexed bridge).
-function sampleResetSteps (proc, n) {
-  const samples = []
-  for (let i = 0; i < n; i++) {
-    proc.reset()
-    proc.next()
-    samples.push(proc.state())
-  }
-  return samples
-}
-
-// Counts Math.log() invocations during fn(), then restores the original — used to verify a
-// precomputed log constant (e.g. this.c.logNoise) isn't being recomputed on every call in a
-// process's _transitionLnPdf hot path.
-function countLogCalls (fn) {
-  const originalLog = Math.log
-  let calls = 0
-  Math.log = (x) => {
-    calls++
-    return originalLog(x)
-  }
-  try {
-    fn()
-  } finally {
-    Math.log = originalLog
-  }
-  return calls
-}
-
-// Asserts the mean per-step transition log-density of a Gaussian-transition process (BM, OU)
-// matches its known theoretical expectation within a CLT tolerance. Each step's residual
-// z = (x_{i+1}-mean)/scale is exactly N(0,1) by construction (_next() draws it directly), so
-// E[-0.5*z^2] = -0.5 exactly and Var(z^2) = 2, giving Var(mean of n such terms) = 1/(2n).
-function assertMeanPerStepLnLMatchesGaussianTransition (proc, path, scale) {
-  const n = path.length - 1
-  const expectedMean = -0.5 - Math.log(scale) - 0.5 * Math.log(2 * Math.PI)
-  const tol = K_SIGMA * Math.sqrt(1 / (2 * n))
-  assert.closeTo(proc.lnL(path) / n, expectedMean, tol)
-}
-
-// Deprecated-alias tests construct through a console.warn-emitting constructor; both silencing
-// it and capturing its message need the same save/restore-on-throw bracketing around the call.
-function withSuppressedWarnings (fn) {
-  const originalWarn = console.warn
-  console.warn = () => {}
-  try {
-    return fn()
-  } finally {
-    console.warn = originalWarn
-  }
-}
-
-function captureWarnings (fn) {
-  const warnings = []
-  const originalWarn = console.warn
-  console.warn = msg => warnings.push(msg)
-  try {
-    fn()
-  } finally {
-    console.warn = originalWarn
-  }
-  return warnings
-}
-
-describe('process._Process', () => {
-  describe('.validate()', () => {
-    it('should throw on undefined parameter', () => {
-      assert.throws(() => Process.validate({ x: undefined }, []), /Required parameters missing/)
-    })
-
-    it('should throw on null parameter', () => {
-      assert.throws(() => Process.validate({ x: null }, []), /Required parameters missing/)
-    })
-
-    it('should throw on NaN parameter', () => {
-      assert.throws(() => Process.validate({ x: NaN }, []), /Required parameters missing/)
-    })
-
-    it('should throw when constraint is violated (<)', () => {
-      assert.throws(() => Process.validate({ x: 5 }, ['x < 3']), /Parameters must satisfy/)
-    })
-
-    it('should throw when constraint is violated (<=)', () => {
-      assert.throws(() => Process.validate({ x: 5 }, ['x <= 3']), /Parameters must satisfy/)
-    })
-
-    it('should throw when constraint is violated (>)', () => {
-      assert.throws(() => Process.validate({ x: 1 }, ['x > 3']), /Parameters must satisfy/)
-    })
-
-    it('should throw when constraint is violated (>=)', () => {
-      assert.throws(() => Process.validate({ x: 1 }, ['x >= 3']), /Parameters must satisfy/)
-    })
-
-    it('should throw when constraint is violated (!=)', () => {
-      assert.throws(() => Process.validate({ x: 3 }, ['x != 3']), /Parameters must satisfy/)
-    })
-
-    it('should not throw when all params are valid and constraints pass', () => {
-      assert.doesNotThrow(() => Process.validate({ mu: 0, sigma: 1 }, ['sigma > 0']))
-    })
-
-    it('should support literal on left side of constraint', () => {
-      assert.doesNotThrow(() => Process.validate({ x: 5 }, ['0 < x']))
-      assert.throws(() => Process.validate({ x: 5 }, ['10 < x']), /Parameters must satisfy/)
-    })
-  })
-})
-
-class RngProcess extends Process {
-  constructor () {
-    super()
-    this.x = 0
-    this.x0 = 0
-  }
-
-  _next () {
-    return this.r.next()
-  }
-}
-
-class StubProcess extends Process {
-  constructor () {
-    super()
-    this.x = 0
-    this.x0 = 0
-  }
-
-  _next () {
-    return this.x + 1
-  }
-}
-
-class BareProcess extends Process {
-  constructor () {
-    super()
-    this.x = 0
-    this.x0 = 0
-  }
-}
-
-class ArrayParamProcess extends Process {
-  constructor (weights) {
-    super()
-    this.p = { weights }
-    this.x = 0
-    this.x0 = 0
-  }
-
-  _next () {
-    return this.x + 1
-  }
-}
-
-describe('process', () => {
-  describe('Process', () => {
-    describe('._next()', () => {
-      it('should throw when not implemented', () => {
-        const p = new BareProcess()
-        assert.throws(() => p.next(), 'Process._next() is not implemented')
-      })
-    })
-
-    describe('.covariogram()', () => {
-      it('should throw when not implemented', () => {
-        const p = new BareProcess()
-        assert.throws(() => p.covariogram(1, 2), 'Process.covariogram() is not implemented')
-      })
-    })
-
-    describe('.mean()', () => {
-      it('should throw when not implemented', () => {
-        const p = new BareProcess()
-        assert.throws(() => p.mean(1), 'Process.mean() is not implemented')
-      })
-    })
-
-    describe('.variance()', () => {
-      it('should throw when not implemented', () => {
-        const p = new BareProcess()
-        assert.throws(() => p.variance(1), 'Process.variance() is not implemented')
-      })
-    })
-
-    describe('.pdf()', () => {
-      it('should throw when not implemented', () => {
-        const p = new BareProcess()
-        assert.throws(() => p.pdf(0, 1), 'Process.pdf() is not implemented')
-      })
-    })
-
-    describe('.marginal()', () => {
-      it('should throw when not implemented', () => {
-        const p = new BareProcess()
-        assert.throws(() => p.marginal(1), 'Process.marginal() is not implemented')
-      })
-    })
-
-    describe('.lnL()', () => {
-      it('should throw when the transition density is not implemented', () => {
-        const p = new BareProcess()
-        assert.throws(() => p.lnL([0, 1]), 'Process._transitionLnPdf() is not implemented')
-      })
-
-      it('should throw when path has fewer than 2 states', () => {
-        const p = new BareProcess()
-        assert.throws(() => p.lnL([0]), /at least 2 states/)
-      })
-
-      it('should throw when path is not an array', () => {
-        const p = new BareProcess()
-        assert.throws(() => p.lnL(null), /at least 2 states/)
-      })
-    })
-
-    describe('.fit()', () => {
-      it('should throw when not implemented', () => {
-        assert.throws(() => BareProcess.fit([0, 1, 2]), 'Process.fit() is not implemented')
-      })
-    })
-
-    describe('.next()', () => {
-      it('should advance state and return the new value', () => {
-        const p = new StubProcess()
-        const s = p.next()
-        assert.strictEqual(s, 1)
-        assert.strictEqual(p.state(), 1)
-      })
-    })
-
-    describe('.path()', () => {
-      it('should return n+1 states starting from initial state', () => {
-        const p = new StubProcess()
-        assert.deepEqual(p.path(5), [0, 1, 2, 3, 4, 5])
-      })
-
-      it('should return 1 state for n=0', () => {
-        const p = new StubProcess()
-        const path = p.path(0)
-        assert.strictEqual(path.length, 1)
-        assert.strictEqual(path[0], 0)
-      })
-
-      it('should not mutate the current state', () => {
-        const p = new StubProcess()
-        p.next()
-        p.path(5)
-        assert.strictEqual(p.state(), 1)
-      })
-
-      it('should advance the PRNG stream', () => {
-        const p1 = new RngProcess()
-        p1.seed(42)
-        p1.path(20)
-        const a1 = p1.next()
-        const p2 = new RngProcess()
-        p2.seed(42)
-        const b1 = p2.next()
-        assert.notStrictEqual(a1, b1)
-      })
-
-      it('should return independent paths on consecutive calls', () => {
-        const p = new RngProcess()
-        p.seed(42)
-        const path1 = p.path(20)
-        const path2 = p.path(20)
-        assert.notDeepEqual(path1, path2)
-      })
-    })
-
-    describe('.ensemble()', () => {
-      it('should return m paths', () => {
-        const p = new RngProcess()
-        p.seed(42)
-        assert.strictEqual(p.ensemble(5, 10).length, 5)
-      })
-
-      it('should return paths of length n+1', () => {
-        const p = new RngProcess()
-        p.seed(42)
-        const e = p.ensemble(3, 7)
-        for (const path of e) {
-          assert.strictEqual(path.length, 8)
-        }
-      })
-
-      it('should return independent paths', () => {
-        const p = new RngProcess()
-        p.seed(42)
-        const e = p.ensemble(3, 20)
-        assert.notDeepEqual(e[0], e[1])
-        assert.notDeepEqual(e[1], e[2])
-      })
-
-      it('should throw for m < 1', () => {
-        const p = new RngProcess()
-        assert.throws(() => p.ensemble(0, 5), /Parameters must satisfy/)
-      })
-
-      it('should throw for n < 1', () => {
-        const p = new RngProcess()
-        assert.throws(() => p.ensemble(3, 0), /Parameters must satisfy/)
-      })
-
-      it('should throw for NaN m', () => {
-        const p = new RngProcess()
-        assert.throws(() => p.ensemble(NaN, 5), /Required parameters missing/)
-      })
-
-      it('should throw for NaN n', () => {
-        const p = new RngProcess()
-        assert.throws(() => p.ensemble(3, NaN), /Required parameters missing/)
-      })
-
-      it('should produce different ensembles on consecutive calls', () => {
-        const p = new RngProcess()
-        p.seed(42)
-        const e1 = p.ensemble(2, 10)
-        const e2 = p.ensemble(2, 10)
-        assert.notDeepEqual(e1, e2)
-      })
-    })
-
-    describe('.reset()', () => {
-      it('should restore the initial state', () => {
-        const p = new StubProcess()
-        p.next()
-        p.next()
-        p.reset()
-        assert.strictEqual(p.state(), 0)
-      })
-    })
-
-    describe('.state()', () => {
-      it('should return the current state', () => {
-        const p = new StubProcess()
-        p.next()
-        assert.strictEqual(p.state(), 1)
-      })
-    })
-
-    describe('.seed()', () => {
-      it('should produce identical paths when seeded identically', () => {
-        const p = new RngProcess()
-        p.seed(42)
-        const path1 = p.path(20)
-        p.seed(42)
-        const path2 = p.path(20)
-        assert.deepEqual(path1, path2)
-      })
-
-      it('should produce different paths for different seeds', () => {
-        const p = new RngProcess()
-        p.seed(1)
-        const path1 = p.path(20)
-        p.reset()
-        p.seed(2)
-        const path2 = p.path(20)
-        assert.notDeepEqual(path1, path2)
-      })
-
-      it('should return this for chaining', () => {
-        const p = new RngProcess()
-        assert.strictEqual(p.seed(0), p)
-      })
-    })
-
-    describe('.params()', () => {
-      it('should return the parameter object', () => {
-        const bm = new BrownianMotion(0.5, 1, 2)
-        assert.deepEqual(bm.params(), { mu: 0.5, sigma: 1, dt: 2 })
-      })
-
-      it('should return an empty object for a process with no parameters', () => {
-        const p = new StubProcess()
-        assert.deepEqual(p.params(), {})
-      })
-
-      it('should return a shallow copy, not the live this.p reference', () => {
-        const bm = new BrownianMotion(0.5, 1, 2)
-        assert.notStrictEqual(bm.params(), bm.p)
-      })
-
-      it('mutating the returned object should not corrupt the instance', () => {
-        const bm = new BrownianMotion(0.5, 1, 2)
-        const p = bm.params()
-        p.mu = 999
-        assert.strictEqual(bm.params().mu, 0.5)
-      })
-
-      it('mutating an array-valued field of the returned object should not corrupt the instance', () => {
-        const ap = new ArrayParamProcess([1, 2, 3])
-        const p = ap.params()
-        p.weights[0] = 999
-        assert.deepEqual(ap.params().weights, [1, 2, 3])
-      })
-    })
-  })
-})
-
-describe('process.BrownianMotion', () => {
-  describe('constructor', () => {
-    it('should throw on sigma = 0', () => {
-      assert.throws(() => new BrownianMotion(0, 0, 1), /Invalid parameters/)
-    })
-
-    it('should throw on sigma < 0', () => {
-      assert.throws(() => new BrownianMotion(0, -1, 1), /Invalid parameters/)
-    })
-
-    it('should throw on dt = 0', () => {
-      assert.throws(() => new BrownianMotion(0, 1, 0), /Invalid parameters/)
-    })
-
-    it('should throw on dt < 0', () => {
-      assert.throws(() => new BrownianMotion(0, 1, -0.5), /Invalid parameters/)
-    })
-
-    it('should throw on mu = NaN', () => {
-      assert.throws(() => new BrownianMotion(NaN, 1, 1), /Invalid parameters/)
-    })
-
-    it('should accept valid parameters', () => {
-      assert.doesNotThrow(() => new BrownianMotion(0, 1, 1))
-      assert.doesNotThrow(() => new BrownianMotion(-2, 0.5, 0.01))
-    })
-
-    it('should start at state 0', () => {
-      const bm = new BrownianMotion(0, 1, 1)
-      assert.strictEqual(bm.state(), 0)
-    })
-  })
-
-  describe('.mean()', () => {
-    it('should return 0 for zero drift at any t', () => {
-      const bm = new BrownianMotion(0, 1, 1)
-      assert.strictEqual(bm.mean(3), 0)
-      assert.strictEqual(bm.mean(100), 0)
-    })
-
-    it('should return 0 at t=0', () => {
-      const bm = new BrownianMotion(1, 1, 1)
-      assert.strictEqual(bm.mean(0), 0)
-    })
-
-    it('should return NaN for t < 0', () => {
-      const bm = new BrownianMotion(0, 1, 1)
-      assert(Number.isNaN(bm.mean(-1)))
-    })
-  })
-
-  describe('.variance()', () => {
-    it('should return 0 at t=0', () => {
-      const bm = new BrownianMotion(0, 1, 1)
-      assert.strictEqual(bm.variance(0), 0)
-    })
-
-    it('should return NaN for t < 0', () => {
-      const bm = new BrownianMotion(0, 1, 1)
-      assert(Number.isNaN(bm.variance(-1)))
-    })
-  })
-
-  describe('.pdf()', () => {
-    it('should return NaN for t = 0', () => {
-      const bm = new BrownianMotion(0, 1, 1)
-      assert(Number.isNaN(bm.pdf(0, 0)))
-    })
-
-    it('should return NaN for t < 0', () => {
-      const bm = new BrownianMotion(0, 1, 1)
-      assert(Number.isNaN(bm.pdf(0, -1)))
-    })
-  })
-
-  describe('.covariogram()', () => {
-    it('should be symmetric', () => {
-      const bm = new BrownianMotion(0, 2, 1)
-      assert.closeTo(bm.covariogram(1, 4), bm.covariogram(4, 1), 1e-10)
-    })
-
-    it('should equal variance at s = t', () => {
-      const bm = new BrownianMotion(0, 2, 1)
-      assert.closeTo(bm.covariogram(3, 3), bm.variance(3), 1e-10)
-    })
-
-    it('should return NaN for s < 0', () => {
-      const bm = new BrownianMotion(0, 1, 1)
-      assert(Number.isNaN(bm.covariogram(-1, 2)))
-    })
-
-    it('should return NaN for t < 0', () => {
-      const bm = new BrownianMotion(0, 1, 1)
-      assert(Number.isNaN(bm.covariogram(2, -1)))
-    })
-  })
-
-  describe('.marginal()', () => {
-    it('should return a Normal distribution with mean and variance matching mean()/variance()', () => {
-      const bm = new BrownianMotion(0.5, 2, 1)
-      const marginal = bm.marginal(3)
-      assert.instanceOf(marginal, Normal)
-      assert.closeTo(marginal.mean(), bm.mean(3), 1e-10)
-      assert.closeTo(marginal.variance(), bm.variance(3), 1e-10)
-    })
-
-    it('should match pdf() at a given point', () => {
-      const bm = new BrownianMotion(0.5, 2, 1)
-      const marginal = bm.marginal(3)
-      assert.closeTo(marginal.pdf(1), bm.pdf(1, 3), 1e-10)
-    })
-
-    it('should round-trip quantile(cdf(x)) = x, exercising the Distribution API beyond pdf/mean/variance', () => {
-      const bm = new BrownianMotion(0.5, 2, 1)
-      const marginal = bm.marginal(3)
-      assert.closeTo(marginal.q(marginal.cdf(1)), 1, 1e-10)
-    })
-
-    it('should throw for t = 0', () => {
-      const bm = new BrownianMotion(0, 1, 1)
-      assert.throws(() => bm.marginal(0), /t must be > 0/)
-    })
-
-    it('should throw for t < 0', () => {
-      const bm = new BrownianMotion(0, 1, 1)
-      assert.throws(() => bm.marginal(-1), /t must be > 0/)
-    })
-  })
-
-  describe('.lnL()', () => {
-    it('should have a mean per-step log-density matching the known transition law (CLT tolerance)', () => {
-      const mu = 0.1
-      const sigma = 1
-      const dt = 1
-      const n = 2000
-      const bm = new BrownianMotion(mu, sigma, dt)
-      bm.seed(42)
-      const path = bm.path(n)
-      assertMeanPerStepLnLMatchesGaussianTransition(bm, path, sigma * Math.sqrt(dt))
-    })
-
-    it('should not recompute log(sigmaDt) per step (this.c.logSigmaDt is precomputed at construction)', () => {
-      const bm = new BrownianMotion(0.3, 1.2, 0.5)
-      const path = [0, 0.4, 1.1, 0.7, 0.2, 0.9, 0.5, 1.3, 0.8, 0.6]
-      const steps = path.length - 1
-      // Only the 2π normalization term is logged per step once sigmaDt's log is cached in
-      // this.c.logSigmaDt; a Math.log(sigmaDt) recomputation per step would double this count.
-      assert.strictEqual(countLogCalls(() => bm.lnL(path)), steps)
-    })
-  })
-
-  describe('.fit()', () => {
-    it('should recover mu and sigma from a long simulated path across seeds', () => {
-      const mu = 0.3
-      const sigma = 1.2
-      const dt = 0.5
-      const n = 20000
-      // CLT tolerance for the exact increment-based MLE: mu_hat ~ N(mu, sigma^2/(n*dt)),
-      // sigma_hat's variance via the delta method on the sample-variance's own CLT tolerance.
-      const tolMu = K_SIGMA * sigma / Math.sqrt(n * dt)
-      const tolSigma = K_SIGMA * sigma * Math.sqrt(1 / (2 * (n - 1)))
-      for (const seed of MOMENT_SEEDS) {
-        const bm = new BrownianMotion(mu, sigma, dt)
-        bm.seed(seed)
-        const fitted = BrownianMotion.fit(bm.path(n), dt)
-        assert.instanceOf(fitted, BrownianMotion)
-        assert.closeTo(fitted.params().mu, mu, tolMu, `seed ${seed}`)
-        assert.closeTo(fitted.params().sigma, sigma, tolSigma, `seed ${seed}`)
-      }
-    })
-
-    it('should default dt to 1', () => {
-      const bm = new BrownianMotion(0.1, 1, 1)
-      bm.seed(1)
-      const fitted = BrownianMotion.fit(bm.path(5000))
-      assert.strictEqual(fitted.params().dt, 1)
-    })
-
-    it('should throw when path has fewer than 3 states', () => {
-      assert.throws(() => BrownianMotion.fit([0, 1], 1), /at least 3 states/)
-    })
-  })
-
-  describe('.path()', () => {
-    it('should have length n+1', () => {
-      const bm = new BrownianMotion(0, 1, 1)
-      assert.strictEqual(bm.path(10).length, 11)
-    })
-
-    it('first element should be initial state 0', () => {
-      const bm = new BrownianMotion(0, 1, 1)
-      assert.strictEqual(bm.path(5)[0], 0)
-    })
-  })
-
-  describe('.reset()', () => {
-    it('should restore initial state to 0', () => {
-      const bm = new BrownianMotion(0, 1, 1)
-      for (let i = 0; i < 10; i++) bm.next()
-      bm.reset()
-      assert.strictEqual(bm.state(), 0)
-    })
-  })
-
-  describe('increments', () => {
-    it('should have mean and variance matching mu*dt and sigma^2*dt across seeds', () => {
-      const mu = 0.1
-      const sigma = 1.5
-      const dt = 0.5
-      const n = 5000
-      // exact rational: BM increment ~ N(mu*dt, sigma^2*dt) by definition of the SDE
-      for (const seed of MOMENT_SEEDS) {
-        const bm = new BrownianMotion(mu, sigma, dt)
-        bm.seed(seed)
-        assertSampleMoments(sampleSteps(bm, n), mu * dt, sigma * sigma * dt, seed)
-      }
-    })
-  })
-})
-
-describe('process.GeometricBrownianMotion', () => {
-  describe('constructor', () => {
-    it('should throw on sigma = 0', () => {
-      assert.throws(() => new GeometricBrownianMotion(0, 0, 1), /Invalid parameters/)
-    })
-
-    it('should throw on sigma < 0', () => {
-      assert.throws(() => new GeometricBrownianMotion(0, -1, 1), /Invalid parameters/)
-    })
-
-    it('should throw on dt = 0', () => {
-      assert.throws(() => new GeometricBrownianMotion(0, 1, 0), /Invalid parameters/)
-    })
-
-    it('should throw on dt < 0', () => {
-      assert.throws(() => new GeometricBrownianMotion(0, 1, -0.5), /Invalid parameters/)
-    })
-
-    it('should throw on mu = NaN', () => {
-      assert.throws(() => new GeometricBrownianMotion(NaN, 1, 1), /Invalid parameters/)
-    })
-
-    it('should accept valid parameters', () => {
-      assert.doesNotThrow(() => new GeometricBrownianMotion(0, 1, 1))
-      assert.doesNotThrow(() => new GeometricBrownianMotion(0.05, 0.2, 0.01))
-    })
-
-    it('should start at state 1', () => {
-      const gbm = new GeometricBrownianMotion(0, 1, 1)
-      assert.strictEqual(gbm.state(), 1)
-    })
-  })
-
-  describe('.path()', () => {
-    it('should have length n+1', () => {
-      const gbm = new GeometricBrownianMotion(0, 1, 1)
-      assert.strictEqual(gbm.path(10).length, 11)
-    })
-
-    it('first element should be 1', () => {
-      const gbm = new GeometricBrownianMotion(0, 1, 1)
-      assert.strictEqual(gbm.path(5)[0], 1)
-    })
-
-    it('all path values should remain positive', () => {
-      const gbm = new GeometricBrownianMotion(0, 1, 1)
-      gbm.seed(42)
-      const path = gbm.path(500)
-      assert(path.every(v => v > 0))
-    })
-  })
-
-  describe('.reset()', () => {
-    it('should restore initial state to 1', () => {
-      const gbm = new GeometricBrownianMotion(0, 1, 1)
-      for (let i = 0; i < 10; i++) gbm.next()
-      gbm.reset()
-      assert.strictEqual(gbm.state(), 1)
-    })
-  })
-
-  describe('.mean()', () => {
-    it('should return 1 at t=0', () => {
-      const gbm = new GeometricBrownianMotion(0.1, 0.2, 1)
-      assert.strictEqual(gbm.mean(0), 1)
-    })
-
-    it('should return NaN for t < 0', () => {
-      const gbm = new GeometricBrownianMotion(0, 1, 1)
-      assert(Number.isNaN(gbm.mean(-1)))
-    })
-
-    it('should be stable after advancing the simulation', () => {
-      const gbm = new GeometricBrownianMotion(0.05, 0.2, 1)
-      const before = gbm.mean(2)
-      for (let i = 0; i < 20; i++) gbm.next()
-      assert.closeTo(gbm.mean(2), before, 1e-10)
-    })
-  })
-
-  describe('.variance()', () => {
-    it('should return 0 at t=0', () => {
-      const gbm = new GeometricBrownianMotion(0.1, 0.2, 1)
-      assert.strictEqual(gbm.variance(0), 0)
-    })
-
-    it('should return NaN for t < 0', () => {
-      const gbm = new GeometricBrownianMotion(0, 1, 1)
-      assert(Number.isNaN(gbm.variance(-1)))
-    })
-  })
-
-  describe('.pdf()', () => {
-    it('should return NaN for t = 0', () => {
-      const gbm = new GeometricBrownianMotion(0, 1, 1)
-      assert(Number.isNaN(gbm.pdf(1, 0)))
-    })
-
-    it('should return NaN for t < 0', () => {
-      const gbm = new GeometricBrownianMotion(0, 1, 1)
-      assert(Number.isNaN(gbm.pdf(1, -1)))
-    })
-
-    it('should return 0 for x = 0 (outside log-normal support)', () => {
-      const gbm = new GeometricBrownianMotion(0, 1, 1)
-      assert.strictEqual(gbm.pdf(0, 1), 0)
-    })
-
-    it('should return 0 for x < 0 (outside log-normal support)', () => {
-      const gbm = new GeometricBrownianMotion(0, 1, 1)
-      assert.strictEqual(gbm.pdf(-1, 1), 0)
-    })
-  })
-
-  describe('.covariogram()', () => {
-    it('should be symmetric', () => {
-      const gbm = new GeometricBrownianMotion(0.05, 0.2, 1)
-      assert.closeTo(gbm.covariogram(1, 4), gbm.covariogram(4, 1), 1e-10)
-    })
-
-    it('should equal variance at s = t', () => {
-      const gbm = new GeometricBrownianMotion(0.05, 0.3, 1)
-      assert.closeTo(gbm.covariogram(2, 2), gbm.variance(2), 1e-10)
-    })
-
-    it('should return NaN for s < 0', () => {
-      const gbm = new GeometricBrownianMotion(0, 1, 1)
-      assert(Number.isNaN(gbm.covariogram(-1, 2)))
-    })
-
-    it('should return NaN for t < 0', () => {
-      const gbm = new GeometricBrownianMotion(0, 1, 1)
-      assert(Number.isNaN(gbm.covariogram(2, -1)))
-    })
-  })
-
-  describe('.marginal()', () => {
-    it('should return a LogNormal distribution matching pdf() at a given point', () => {
-      const gbm = new GeometricBrownianMotion(0.05, 0.2, 1)
-      const marginal = gbm.marginal(2)
-      assert.instanceOf(marginal, LogNormal)
-      assert.closeTo(marginal.pdf(1.5), gbm.pdf(1.5, 2), 1e-10)
-    })
-
-    it('should have mean and variance matching mean()/variance()', () => {
-      const gbm = new GeometricBrownianMotion(0.05, 0.2, 1)
-      const marginal = gbm.marginal(2)
-      assert.closeTo(marginal.mean(), gbm.mean(2), 1e-10)
-      assert.closeTo(marginal.variance(), gbm.variance(2), 1e-10)
-    })
-
-    it('should round-trip quantile(cdf(x)) = x, exercising the Distribution API beyond pdf/mean/variance', () => {
-      const gbm = new GeometricBrownianMotion(0.05, 0.2, 1)
-      const marginal = gbm.marginal(2)
-      assert.closeTo(marginal.q(marginal.cdf(1.5)), 1.5, 1e-10)
-    })
-
-    it('should throw for t = 0', () => {
-      const gbm = new GeometricBrownianMotion(0, 1, 1)
-      assert.throws(() => gbm.marginal(0), /t must be > 0/)
-    })
-
-    it('should throw for t < 0', () => {
-      const gbm = new GeometricBrownianMotion(0, 1, 1)
-      assert.throws(() => gbm.marginal(-1), /t must be > 0/)
-    })
-  })
-
-  describe('.lnL()', () => {
-    it('should return -Infinity, not throw, when the path visits a non-positive state', () => {
-      const gbm = new GeometricBrownianMotion(0.1, 0.25, 0.5)
-      assert.strictEqual(gbm.lnL([1, -1]), -Infinity)
-    })
-
-    it('should return -Infinity, not NaN, when the non-positive state is not the last one', () => {
-      const gbm = new GeometricBrownianMotion(0.1, 0.25, 0.5)
-      assert.strictEqual(gbm.lnL([1, -1, 2]), -Infinity)
-    })
-
-    it('should match the known transition law plus the observed path Jacobian (CLT tolerance)', () => {
-      const mu = 0.05
-      const sigma = 0.2
-      const dt = 1
-      const n = 2000
-      const gbm = new GeometricBrownianMotion(mu, sigma, dt)
-      gbm.seed(42)
-      const path = gbm.path(n)
-      const noise = sigma * Math.sqrt(dt)
-      // Each step's residual z = (log(x_{i+1}/x_i)-drift)/noise is exactly N(0,1) by construction,
-      // so E[-0.5*z^2] = -0.5 exactly (Var(z^2) = 2). The -log(x_{i+1}) Jacobian term is not part
-      // of that randomness — it's summed directly from the actually observed path values.
-      let logJacobianSum = 0
-      for (let i = 0; i < n; i++) logJacobianSum += Math.log(path[i + 1])
-      const expectedTotal = n * (-0.5 - Math.log(noise) - 0.5 * Math.log(2 * Math.PI)) - logJacobianSum
-      const tol = K_SIGMA * Math.sqrt(n / 2)
-      assert.closeTo(gbm.lnL(path), expectedTotal, tol)
-    })
-
-    it('should not recompute log(noise) per step (this.c.logNoise is precomputed at construction)', () => {
-      const gbm = new GeometricBrownianMotion(0.1, 0.25, 0.5)
-      const path = [1.0, 1.2, 0.9, 1.5, 1.1, 1.3, 0.95, 1.4, 1.05, 1.25]
-      const steps = path.length - 1
-      // Each step already logs log(xNext/xPrev), log(2π), and log(xNext) (the Jacobian term);
-      // a Math.log(noise) recomputation per step would push this to 4 calls/step instead of 3.
-      assert.strictEqual(countLogCalls(() => gbm.lnL(path)), 3 * steps)
-    })
-  })
-
-  describe('.fit()', () => {
-    it('should recover mu and sigma from a long simulated path across seeds', () => {
-      const mu = 0.05
-      const sigma = 0.3
-      const dt = 0.5
-      const n = 20000
-      // Same CLT-derived tolerance style as BrownianMotion.fit(), applied to log-returns
-      // (which are i.i.d. N((mu-sigma^2/2)*dt, sigma^2*dt), same as BM increments).
-      const tolSigma = K_SIGMA * sigma * Math.sqrt(1 / (2 * (n - 1)))
-      // mu_hat = mean(logReturns)/dt + sigma_hat^2/2, so its tolerance also absorbs sigma_hat's error
-      const tolMu = K_SIGMA * sigma / Math.sqrt(n * dt) + tolSigma * sigma
-      for (const seed of MOMENT_SEEDS) {
-        const gbm = new GeometricBrownianMotion(mu, sigma, dt)
-        gbm.seed(seed)
-        const fitted = GeometricBrownianMotion.fit(gbm.path(n), dt)
-        assert.instanceOf(fitted, GeometricBrownianMotion)
-        assert.closeTo(fitted.params().sigma, sigma, tolSigma, `seed ${seed}`)
-        assert.closeTo(fitted.params().mu, mu, tolMu, `seed ${seed}`)
-      }
-    })
-
-    it('should default dt to 1', () => {
-      const gbm = new GeometricBrownianMotion(0.05, 0.2, 1)
-      gbm.seed(1)
-      const fitted = GeometricBrownianMotion.fit(gbm.path(5000))
-      assert.strictEqual(fitted.params().dt, 1)
-    })
-
-    it('should throw when path has fewer than 3 states', () => {
-      assert.throws(() => GeometricBrownianMotion.fit([1, 1.1], 1), /at least 3 states/)
-    })
-
-    it('should throw when path contains a non-positive state', () => {
-      assert.throws(() => GeometricBrownianMotion.fit([1, 1.1, 0, 1.2], 1), /positive states/)
-    })
-  })
-
-  describe('log-returns', () => {
-    it('should have mean and variance matching the GBM/Itô log-return identity across seeds', () => {
-      const mu = 0.05
-      const sigma = 0.2
-      const dt = 1
-      const n = 5000
-      // exact rational: log(X_{t+dt}/X_t) ~ N((mu - sigma^2/2)*dt, sigma^2*dt) by Itô's lemma,
-      // independent of the sampler implementation
-      const meanLR = (mu - 0.5 * sigma * sigma) * dt
-      const varLR = sigma * sigma * dt
-      for (const seed of MOMENT_SEEDS) {
-        const gbm = new GeometricBrownianMotion(mu, sigma, dt)
-        gbm.seed(seed)
-        const logReturns = sampleSteps(gbm, n, (curr, prev) => Math.log(curr / prev))
-        assertSampleMoments(logReturns, meanLR, varLR, seed)
-      }
-    })
-  })
-})
-
-describe('process.OrnsteinUhlenbeck', () => {
-  describe('constructor', () => {
-    it('should throw on theta = 0', () => {
-      assert.throws(() => new OrnsteinUhlenbeck(0, 0, 1, 1), /Invalid parameters/)
-    })
-
-    it('should throw on theta < 0', () => {
-      assert.throws(() => new OrnsteinUhlenbeck(-1, 0, 1, 1), /Invalid parameters/)
-    })
-
-    it('should throw on sigma = 0', () => {
-      assert.throws(() => new OrnsteinUhlenbeck(1, 0, 0, 1), /Invalid parameters/)
-    })
-
-    it('should throw on sigma < 0', () => {
-      assert.throws(() => new OrnsteinUhlenbeck(1, 0, -1, 1), /Invalid parameters/)
-    })
-
-    it('should throw on dt = 0', () => {
-      assert.throws(() => new OrnsteinUhlenbeck(1, 0, 1, 0), /Invalid parameters/)
-    })
-
-    it('should throw on dt < 0', () => {
-      assert.throws(() => new OrnsteinUhlenbeck(1, 0, 1, -0.5), /Invalid parameters/)
-    })
-
-    it('should throw on mu = NaN', () => {
-      assert.throws(() => new OrnsteinUhlenbeck(1, NaN, 1, 1), /Invalid parameters/)
-    })
-
-    it('should accept valid parameters', () => {
-      assert.doesNotThrow(() => new OrnsteinUhlenbeck(2, 1, 0.5, 0.1))
-    })
-
-    it('should start at state 0', () => {
-      const ou = new OrnsteinUhlenbeck(1, 2, 0.5, 0.1)
-      assert.strictEqual(ou.state(), 0)
-    })
-  })
-
-  describe('.mean()', () => {
-    it('should return 0 at t=0', () => {
-      const ou = new OrnsteinUhlenbeck(1, 5, 1, 0.1)
-      assert.strictEqual(ou.mean(0), 0)
-    })
-
-    it('should return NaN for t < 0', () => {
-      const ou = new OrnsteinUhlenbeck(1, 0, 1, 1)
-      assert(isNaN(ou.mean(-1)))
-    })
-
-    it('should be stable after advancing the simulation', () => {
-      const ou = new OrnsteinUhlenbeck(2, 3, 1, 0.1)
-      const before = ou.mean(1)
-      for (let i = 0; i < 20; i++) ou.next()
-      assert.closeTo(ou.mean(1), before, 1e-10)
-    })
-  })
-
-  describe('.variance()', () => {
-    it('should return 0 at t=0', () => {
-      const ou = new OrnsteinUhlenbeck(1, 0, 1, 1)
-      assert.strictEqual(ou.variance(0), 0)
-    })
-
-    it('should return NaN for t < 0', () => {
-      const ou = new OrnsteinUhlenbeck(1, 0, 1, 1)
-      assert(isNaN(ou.variance(-1)))
-    })
-  })
-
-  describe('.pdf()', () => {
-    it('should return NaN for t = 0', () => {
-      const ou = new OrnsteinUhlenbeck(1, 0, 1, 1)
-      assert(Number.isNaN(ou.pdf(0, 0)))
-    })
-
-    it('should return NaN for t < 0', () => {
-      const ou = new OrnsteinUhlenbeck(1, 0, 1, 1)
-      assert(Number.isNaN(ou.pdf(0, -1)))
-    })
-  })
-
-  describe('.covariogram()', () => {
-    it('should be symmetric', () => {
-      const ou = new OrnsteinUhlenbeck(2, 0, 0.5, 0.1)
-      assert.closeTo(ou.covariogram(1, 3), ou.covariogram(3, 1), 1e-10)
-    })
-
-    it('should equal variance at s = t', () => {
-      const ou = new OrnsteinUhlenbeck(2, 0, 0.5, 0.1)
-      assert.closeTo(ou.covariogram(2, 2), ou.variance(2), 1e-10)
-    })
-
-    it('should return NaN for s < 0', () => {
-      const ou = new OrnsteinUhlenbeck(1, 0, 1, 1)
-      assert(Number.isNaN(ou.covariogram(-1, 2)))
-    })
-
-    it('should return NaN for t < 0', () => {
-      const ou = new OrnsteinUhlenbeck(1, 0, 1, 1)
-      assert(Number.isNaN(ou.covariogram(2, -1)))
-    })
-  })
-
-  describe('.marginal()', () => {
-    it('should return a Normal distribution with mean and variance matching mean()/variance()', () => {
-      const ou = new OrnsteinUhlenbeck(0.5, 3, 2, 0.1)
-      const marginal = ou.marginal(2)
-      assert.instanceOf(marginal, Normal)
-      assert.closeTo(marginal.mean(), ou.mean(2), 1e-10)
-      assert.closeTo(marginal.variance(), ou.variance(2), 1e-10)
-    })
-
-    it('should match pdf() at a given point', () => {
-      const ou = new OrnsteinUhlenbeck(0.5, 3, 2, 0.1)
-      const marginal = ou.marginal(2)
-      assert.closeTo(marginal.pdf(2), ou.pdf(2, 2), 1e-10)
-    })
-
-    it('should round-trip quantile(cdf(x)) = x, exercising the Distribution API beyond pdf/mean/variance', () => {
-      const ou = new OrnsteinUhlenbeck(0.5, 3, 2, 0.1)
-      const marginal = ou.marginal(2)
-      assert.closeTo(marginal.q(marginal.cdf(2)), 2, 1e-10)
-    })
-
-    it('should throw for t = 0', () => {
-      const ou = new OrnsteinUhlenbeck(1, 0, 1, 1)
-      assert.throws(() => ou.marginal(0), /t must be > 0/)
-    })
-
-    it('should throw for t < 0', () => {
-      const ou = new OrnsteinUhlenbeck(1, 0, 1, 1)
-      assert.throws(() => ou.marginal(-1), /t must be > 0/)
-    })
-  })
-
-  describe('.lnL()', () => {
-    it('should have a mean per-step log-density matching the known transition law (CLT tolerance)', () => {
-      const theta = 0.5
-      const mu = 3
-      const sigma = 2
-      const dt = 0.1
-      const n = 2000
-      const ou = new OrnsteinUhlenbeck(theta, mu, sigma, dt)
-      ou.seed(42)
-      const path = ou.path(n)
-      const decay = Math.exp(-theta * dt)
-      const noise = sigma * Math.sqrt((1 - decay * decay) / (2 * theta))
-      assertMeanPerStepLnLMatchesGaussianTransition(ou, path, noise)
-    })
-
-    it('should not recompute log(noise) per step (this.c.logNoise is precomputed at construction)', () => {
-      const ou = new OrnsteinUhlenbeck(0.8, 0.5, 0.6, 0.25)
-      const path = [1.0, 0.9, 0.6, 0.8, 0.7, 0.95, 0.65, 0.85, 0.75, 0.9]
-      const steps = path.length - 1
-      // Only the 2π normalization term is logged per step once noise's log is cached in
-      // this.c.logNoise; a Math.log(noise) recomputation per step would double this count.
-      assert.strictEqual(countLogCalls(() => ou.lnL(path)), steps)
-    })
-  })
-
-  describe('.fit()', () => {
-    it('should recover theta, mu, and sigma from a long simulated path across seeds', () => {
-      const theta = 0.8
-      const mu = 2
-      const sigma = 1.5
-      const dt = 0.2
-      const n = 30000
-      // fit() is OLS on the exact AR(1) transition y = a + b*x + eps, b = exp(-theta*dt),
-      // a = mu*(1-b), eps ~ N(0, s2) with s2 = sigma^2*(1-b^2)/(2*theta). Tolerances below are
-      // the classical AR(1)-OLS asymptotic standard errors (Hamilton, "Time Series Analysis",
-      // eq. 8.2.16: sqrt(n)(bHat-b) -> N(0, 1-b^2)) propagated through theta/mu/sigma's
-      // back-substitution formulas via the delta method, replacing the old flat 15% band
-      // (which was loose enough to pass even with a systematic back-substitution bug — see
-      // solutions/testing/2026-07-28-1601-flat-tolerance-masks-estimator-bugs.md).
-      const b = Math.exp(-theta * dt)
-      const s2 = sigma * sigma * (1 - b * b) / (2 * theta)
-      const varB = (1 - b * b) / n
-      // theta = -ln(b)/dt, d(theta)/d(b) = -1/(b*dt)
-      const varTheta = varB / (b * b * dt * dt)
-      // mu = a/(1-b); propagating Var(a), Var(b), and Cov(a,b) from OLS through the delta
-      // method makes every mu^2 term cancel, leaving this parameter-free closed form.
-      const varMu = s2 / (n * (1 - b) * (1 - b))
-      // sigma^2 = s2*2*theta/(1-b^2) = s2*(-2*ln(b)/dt)/(1-b^2); by Cochran's theorem s2 and b
-      // are exactly independent (conditional on the path) for Gaussian OLS, so their variance
-      // contributions add without a covariance cross-term. varS2 uses the same chi-squared-type
-      // sampling-variance form as assertSampleMoments's tolVariance (expected*sqrt(2/(n-1))).
-      const dSigma2dS2 = 2 * theta / (1 - b * b)
-      const dSigma2dB = -(2 * s2 / dt) * ((1 - b * b) / b + 2 * b * Math.log(b)) / ((1 - b * b) * (1 - b * b))
-      const varS2 = 2 * s2 * s2 / n
-      const varSigma2 = dSigma2dS2 * dSigma2dS2 * varS2 + dSigma2dB * dSigma2dB * varB
-      const varSigma = varSigma2 / (4 * sigma * sigma)
-      const tolTheta = K_SIGMA * Math.sqrt(varTheta)
-      const tolMu = K_SIGMA * Math.sqrt(varMu)
-      const tolSigma = K_SIGMA * Math.sqrt(varSigma)
-      for (const seed of MOMENT_SEEDS) {
-        const ou = new OrnsteinUhlenbeck(theta, mu, sigma, dt)
-        ou.seed(seed)
-        const fitted = OrnsteinUhlenbeck.fit(ou.path(n), dt)
-        assert.instanceOf(fitted, OrnsteinUhlenbeck)
-        assert.closeTo(fitted.params().theta, theta, tolTheta, `seed ${seed}: theta`)
-        assert.closeTo(fitted.params().mu, mu, tolMu, `seed ${seed}: mu`)
-        assert.closeTo(fitted.params().sigma, sigma, tolSigma, `seed ${seed}: sigma`)
-      }
-    })
-
-    it('should default dt to 1', () => {
-      const ou = new OrnsteinUhlenbeck(0.5, 1, 1, 1)
-      ou.seed(1)
-      const fitted = OrnsteinUhlenbeck.fit(ou.path(20000))
-      assert.strictEqual(fitted.params().dt, 1)
-    })
-
-    it('should throw when path has fewer than 4 states', () => {
-      assert.throws(() => OrnsteinUhlenbeck.fit([0, 1, 2], 1), /at least 4 states/)
-    })
-
-    it('should throw when the AR(1) slope estimate is out of (0,1)', () => {
-      // A perfectly linear path drives the OLS slope to exactly 1 (X_{n+1}-X_n constant),
-      // outside the valid exp(-theta*dt) in (0,1) range for any theta, dt > 0.
-      assert.throws(() => OrnsteinUhlenbeck.fit([0, 1, 2, 3, 4, 5], 1), /AR\(1\) slope/)
-    })
-  })
-
-  describe('.reset()', () => {
-    it('should restore initial state to 0', () => {
-      const ou = new OrnsteinUhlenbeck(1, 2, 0.5, 0.1)
-      for (let i = 0; i < 10; i++) ou.next()
-      ou.reset()
-      assert.strictEqual(ou.state(), 0)
-    })
-  })
-
-  describe('stationarity', () => {
-    it('should converge to stationary mean mu and variance sigma^2/(2*theta) across seeds', () => {
-      const theta = 2; const mu = 3; const sigma = 1; const dt = 0.1
-      // exact rational: OU stationary distribution is N(mu, sigma^2/(2*theta))
-      const stationaryVariance = sigma * sigma / (2 * theta)
-      for (const seed of MOMENT_SEEDS) {
-        const ou = new OrnsteinUhlenbeck(theta, mu, sigma, dt)
-        ou.seed(seed)
-        for (let i = 0; i < 500; i++) ou.next()
-        // lag-1 autocorrelation is exp(-theta*dt)=exp(-0.2)≈0.82; thin by 20 to get
-        // independent draws (lag-20 autocorrelation ≈ 0.018)
-        const samples = []
-        for (let i = 0; i < 20000; i++) {
-          ou.next()
-          if (i % 20 === 0) samples.push(ou.state())
-        }
-        assertSampleMoments(samples, mu, stationaryVariance, seed)
-      }
-    })
-  })
-})
-
-describe('process.BrownianBridge', () => {
-  describe('constructor', () => {
-    it('should throw on sigma = 0', () => {
-      assert.throws(() => new BrownianBridge(0, 1, 0.1), /Invalid parameters/)
-    })
-
-    it('should throw on sigma < 0', () => {
-      assert.throws(() => new BrownianBridge(-1, 1, 0.1), /Invalid parameters/)
-    })
-
-    it('should throw on T = 0', () => {
-      assert.throws(() => new BrownianBridge(1, 0, 0.1), /Invalid parameters/)
-    })
-
-    it('should throw on T < 0', () => {
-      assert.throws(() => new BrownianBridge(1, -1, 0.1), /Invalid parameters/)
-    })
-
-    it('should throw on dt = 0', () => {
-      assert.throws(() => new BrownianBridge(1, 1, 0), /Invalid parameters/)
-    })
-
-    it('should throw on dt < 0', () => {
-      assert.throws(() => new BrownianBridge(1, 1, -0.1), /Invalid parameters/)
-    })
-
-    it('should throw on sigma = NaN', () => {
-      assert.throws(() => new BrownianBridge(NaN, 1, 0.1), /Invalid parameters/)
-    })
-
-    it('should accept valid parameters', () => {
-      assert.doesNotThrow(() => new BrownianBridge(1, 1, 0.1))
-      assert.doesNotThrow(() => new BrownianBridge(0.5, 2, 0.5))
-    })
-
-    it('should start at state 0', () => {
-      const bb = new BrownianBridge(1, 1, 0.1)
-      assert.strictEqual(bb.state(), 0)
-    })
-  })
-
-  describe('terminal value', () => {
-    it('should return 0 exactly at terminal step N = T/dt', () => {
-      const T = 1
-      const dt = 0.1
-      const N = Math.round(T / dt)
-      const bb = new BrownianBridge(1, T, dt)
-      bb.seed(42)
-      for (let i = 0; i < N; i++) bb.next()
-      assert.strictEqual(bb.state(), 0)
-    })
-
-    it('should stay at 0 after terminal time', () => {
-      const bb = new BrownianBridge(1, 1, 0.1)
-      bb.seed(42)
-      const N = 10
-      for (let i = 0; i < N + 5; i++) bb.next()
-      assert.strictEqual(bb.state(), 0)
-    })
-  })
-
-  describe('.reset()', () => {
-    it('should restore initial state and time index', () => {
-      const T = 1
-      const dt = 0.1
-      const N = Math.round(T / dt)
-      const bb = new BrownianBridge(1, T, dt)
-      bb.seed(42)
-      for (let i = 0; i < N; i++) bb.next()
-      bb.reset()
-      assert.strictEqual(bb.state(), 0)
-      for (let i = 0; i < N; i++) bb.next()
-      assert.strictEqual(bb.state(), 0)
-    })
-  })
-
-  describe('.path()', () => {
-    it('should return N+1 states starting from 0', () => {
-      const bb = new BrownianBridge(1, 1, 0.1)
-      const path = bb.path(10)
-      assert.strictEqual(path.length, 11)
-      assert.strictEqual(path[0], 0)
-    })
-
-    it('should end at 0 at the terminal step', () => {
-      const T = 1
-      const dt = 0.1
-      const N = Math.round(T / dt)
-      const bb = new BrownianBridge(1, T, dt)
-      bb.seed(42)
-      const path = bb.path(N)
-      assert.strictEqual(path[N], 0)
-    })
-
-    it('should not mutate the current state or time index', () => {
-      const T = 1
-      const dt = 0.1
-      const N = Math.round(T / dt)
-      const bb = new BrownianBridge(1, T, dt)
-      bb.seed(42)
-      for (let i = 0; i < 5; i++) bb.next()
-      const stateBefore = bb.state()
-      bb.path(N)
-      assert.strictEqual(bb.state(), stateBefore)
-      // Remaining 5 steps should still reach 0
-      for (let i = 0; i < 5; i++) bb.next()
-      assert.strictEqual(bb.state(), 0)
-    })
-  })
-
-  describe('.mean()', () => {
-    it('should return 0 for t >= 0', () => {
-      const bb = new BrownianBridge(1, 1, 0.1)
-      assert.strictEqual(bb.mean(0), 0)
-      assert.strictEqual(bb.mean(0.5), 0)
-      assert.strictEqual(bb.mean(1), 0)
-      assert.strictEqual(bb.mean(2), 0)
-    })
-
-    it('should return NaN for t < 0', () => {
-      const bb = new BrownianBridge(1, 1, 0.1)
-      assert(Number.isNaN(bb.mean(-1)))
-    })
-  })
-
-  describe('.variance()', () => {
-    it('should return 0 at t=0', () => {
-      const bb = new BrownianBridge(1, 1, 0.1)
-      assert.strictEqual(bb.variance(0), 0)
-    })
-
-    it('should return 0 at t=T', () => {
-      const bb = new BrownianBridge(1, 1, 0.1)
-      assert.strictEqual(bb.variance(1), 0)
-    })
-
-    it('should return 0 for t > T', () => {
-      const bb = new BrownianBridge(1, 1, 0.1)
-      assert.strictEqual(bb.variance(2), 0)
-    })
-
-    it('should return NaN for t < 0', () => {
-      const bb = new BrownianBridge(1, 1, 0.1)
-      assert(Number.isNaN(bb.variance(-1)))
-    })
-  })
-
-  describe('.covariogram()', () => {
-    it('should be symmetric', () => {
-      const bb = new BrownianBridge(1, 2, 0.1)
-      assert.closeTo(bb.covariogram(0.5, 1.5), bb.covariogram(1.5, 0.5), 1e-10)
-    })
-
-    it('should equal variance at s = t', () => {
-      const sigma = 2; const T = 1
-      const bb = new BrownianBridge(sigma, T, 0.1)
-      assert.closeTo(bb.covariogram(0.4, 0.4), bb.variance(0.4), 1e-10)
-    })
-
-    it('should return 0 for s > T', () => {
-      const bb = new BrownianBridge(1, 1, 0.1)
-      assert.strictEqual(bb.covariogram(1.5, 0.5), 0)
-    })
-
-    it('should return 0 for t > T', () => {
-      const bb = new BrownianBridge(1, 1, 0.1)
-      assert.strictEqual(bb.covariogram(0.5, 1.5), 0)
-    })
-
-    it('should return 0 at t = T', () => {
-      const bb = new BrownianBridge(1, 1, 0.1)
-      assert.strictEqual(bb.covariogram(0.5, 1), 0)
-    })
-
-    it('should return NaN for s < 0', () => {
-      const bb = new BrownianBridge(1, 1, 0.1)
-      assert(Number.isNaN(bb.covariogram(-0.5, 0.5)))
-    })
-
-    it('should return NaN for t < 0', () => {
-      const bb = new BrownianBridge(1, 1, 0.1)
-      assert(Number.isNaN(bb.covariogram(0.5, -0.5)))
-    })
-  })
-
-  describe('.pdf()', () => {
-    it('should return NaN for t < 0', () => {
-      const bb = new BrownianBridge(1, 2, 0.1)
-      assert(Number.isNaN(bb.pdf(0, -1)))
-    })
-
-    it('should return Infinity at x=0 when t=0 (point mass at origin)', () => {
-      const bb = new BrownianBridge(1, 2, 0.1)
-      assert.strictEqual(bb.pdf(0, 0), Infinity)
-    })
-
-    it('should return 0 at x≠0 when t=0', () => {
-      const bb = new BrownianBridge(1, 2, 0.1)
-      assert.strictEqual(bb.pdf(1, 0), 0)
-    })
-
-    it('should return Infinity at x=0 when t=T (pinned to 0)', () => {
-      const bb = new BrownianBridge(1, 2, 0.1)
-      assert.strictEqual(bb.pdf(0, 2), Infinity)
-    })
-
-    it('should return 0 at x≠0 when t=T', () => {
-      const bb = new BrownianBridge(1, 2, 0.1)
-      assert.strictEqual(bb.pdf(1, 2), 0)
-    })
-
-    it('should return Infinity at x=0 for t > T', () => {
-      const bb = new BrownianBridge(1, 2, 0.1)
-      assert.strictEqual(bb.pdf(0, 3), Infinity)
-    })
-
-    it('should return 0 at x≠0 for t > T', () => {
-      const bb = new BrownianBridge(1, 2, 0.1)
-      assert.strictEqual(bb.pdf(1, 3), 0)
-    })
-
-    it('should be symmetric around 0 (pdf(-x, t) = pdf(x, t))', () => {
-      const bb = new BrownianBridge(1, 2, 0.1)
-      assert.closeTo(bb.pdf(-1, 1), bb.pdf(1, 1), 1e-10)
-    })
-  })
-
-  describe('.marginal()', () => {
-    it('should return a Normal distribution with variance matching variance()', () => {
-      const bb = new BrownianBridge(2, 4, 0.1)
-      const marginal = bb.marginal(2)
-      assert.instanceOf(marginal, Normal)
-      assert.strictEqual(marginal.mean(), 0)
-      assert.closeTo(marginal.variance(), bb.variance(2), 1e-10)
-    })
-
-    it('should match pdf() at a given point', () => {
-      const bb = new BrownianBridge(1, 2, 0.1)
-      const marginal = bb.marginal(1)
-      assert.closeTo(marginal.pdf(1), bb.pdf(1, 1), 1e-10)
-    })
-
-    it('should round-trip quantile(cdf(x)) = x, exercising the Distribution API beyond pdf/mean/variance', () => {
-      const bb = new BrownianBridge(1, 2, 0.1)
-      const marginal = bb.marginal(1)
-      assert.closeTo(marginal.q(marginal.cdf(1)), 1, 1e-10)
-    })
-
-    it('should throw for t = 0', () => {
-      const bb = new BrownianBridge(1, 2, 0.1)
-      assert.throws(() => bb.marginal(0), /0 < t < T/)
-    })
-
-    it('should throw for t = T', () => {
-      const bb = new BrownianBridge(1, 2, 0.1)
-      assert.throws(() => bb.marginal(2), /0 < t < T/)
-    })
-
-    it('should throw for t > T', () => {
-      const bb = new BrownianBridge(1, 2, 0.1)
-      assert.throws(() => bb.marginal(3), /0 < t < T/)
-    })
-  })
-
-  describe('increments', () => {
-    it('should have mean and variance matching sigma^2*dt*(T-dt)/T at t=0 across seeds', () => {
-      const sigma = 1.5
-      const T = 100
-      const dt = 1
-      const n = 5000
-      // exact rational: standard Brownian bridge variance formula sigma^2*t*(T-t)/T evaluated at t=dt
-      const expectedVariance = sigma * sigma * dt * (T - dt) / T
-      for (const seed of MOMENT_SEEDS) {
-        const bb = new BrownianBridge(sigma, T, dt)
-        bb.seed(seed)
-        assertSampleMoments(sampleResetSteps(bb, n), 0, expectedVariance, seed)
-      }
-    })
-  })
-
-  describe('.fit()', () => {
-    it('should recover sigma from a long simulated path across seeds', () => {
-      const sigma = 1.4
-      const T = 2000
-      const dt = 1
-      const N = T / dt
-      // Every step's conditional draw is exact (no discretization error), so the MLE's
-      // sigma^2 estimate is exactly sigma^2/(N-1) times a chi-square(N-1) variate; propagating
-      // its variance (2*sigma^4/(N-1)) through the delta method gives sigma_hat's own tolerance.
-      const tolSigma = K_SIGMA * sigma * Math.sqrt(1 / (2 * (N - 1)))
-      for (const seed of MOMENT_SEEDS) {
-        const bb = new BrownianBridge(sigma, T, dt)
-        bb.seed(seed)
-        const fitted = BrownianBridge.fit(bb.path(N), T, dt)
-        assert.instanceOf(fitted, BrownianBridge)
-        assert.closeTo(fitted.params().sigma, sigma, tolSigma, `seed ${seed}`)
-      }
-    })
-
-    it('should default dt to 0.1', () => {
-      const bb = new BrownianBridge(1, 10, 0.1)
-      bb.seed(1)
-      const fitted = BrownianBridge.fit(bb.path(100), 10)
-      assert.strictEqual(fitted.params().dt, 0.1)
-    })
-
-    it('should throw when T is not > 0', () => {
-      assert.throws(() => BrownianBridge.fit([0, 0], -1, 1), /Invalid parameters/)
-    })
-
-    it('should throw when dt is not > 0', () => {
-      assert.throws(() => BrownianBridge.fit([0, 0], 10, 0), /Invalid parameters/)
-    })
-
-    it('should throw when T/dt is less than 2', () => {
-      assert.throws(() => BrownianBridge.fit([0, 0], 1, 1), /at least 2/)
-    })
-
-    it('should succeed at exactly T/dt = 2 (the minimum estimable case)', () => {
-      const sigma = 1.5
-      const T = 1
-      const dt = 0.5
-      const bb = new BrownianBridge(sigma, T, dt)
-      bb.seed(1)
-      const fitted = BrownianBridge.fit(bb.path(2), T, dt)
-      assert.instanceOf(fitted, BrownianBridge)
-      assert.isAbove(fitted.params().sigma, 0)
-    })
-
-    it('should throw when path is not an array', () => {
-      assert.throws(() => BrownianBridge.fit(null, 10, 1), /exactly T\/dt \+ 1 states/)
-    })
-
-    it('should throw when path length does not match T/dt + 1', () => {
-      assert.throws(() => BrownianBridge.fit([0, 1, 0], 4, 1), /exactly T\/dt \+ 1 states/)
-    })
-  })
-})
-
-describe('process.AR1', () => {
-  describe('constructor', () => {
-    it('should throw on sigma = 0', () => {
-      assert.throws(() => new AR1(0.5, 0), /Invalid parameters/)
-    })
-
-    it('should throw on sigma < 0', () => {
-      assert.throws(() => new AR1(0.5, -1), /Invalid parameters/)
-    })
-
-    it('should throw on sigma = NaN', () => {
-      assert.throws(() => new AR1(0.5, NaN), /Invalid parameters/)
-    })
-
-    it('should throw on phi = NaN', () => {
-      assert.throws(() => new AR1(NaN, 1), /Invalid parameters/)
-    })
-
-    it('should accept |phi| >= 1 without throwing (non-stationary)', () => {
-      assert.doesNotThrow(() => new AR1(1, 1))
-      assert.doesNotThrow(() => new AR1(1.5, 1))
-      assert.doesNotThrow(() => new AR1(-1.2, 1))
-    })
-
-    it('should accept valid stationary parameters', () => {
-      assert.doesNotThrow(() => new AR1(0.5, 1))
-      assert.doesNotThrow(() => new AR1(-0.9, 2))
-    })
-
-    it('should start at state 0', () => {
-      assert.strictEqual(new AR1(0.5, 1).state(), 0)
-    })
-  })
-
-  describe('.mean()', () => {
-    it('should return 0 for t >= 0', () => {
-      const ar1 = new AR1(0.5, 1)
-      assert.strictEqual(ar1.mean(0), 0)
-      assert.strictEqual(ar1.mean(5), 0)
-      assert.strictEqual(ar1.mean(100), 0)
-    })
-
-    it('should return NaN for t < 0', () => {
-      const ar1 = new AR1(0.5, 1)
-      assert(Number.isNaN(ar1.mean(-1)))
-    })
-  })
-
-  describe('.variance()', () => {
-    it('should return 0 at t = 0', () => {
-      const ar1 = new AR1(0.5, 1)
-      assert.strictEqual(ar1.variance(0), 0)
-    })
-
-    it('should return 0 at t = 0 for phi = 0 (phi2 = 0 would otherwise hit 0*log(0) = NaN)', () => {
-      const ar1 = new AR1(0, 1)
-      assert.strictEqual(ar1.variance(0), 0)
-    })
-
-    it('should return NaN for t < 0', () => {
-      const ar1 = new AR1(0.5, 1)
-      assert(Number.isNaN(ar1.variance(-1)))
-    })
-
-    it('should grow monotonically for |phi| > 1', () => {
-      const ar1 = new AR1(1.5, 1)
-      assert(ar1.variance(10) > ar1.variance(5))
-      assert(ar1.variance(20) > ar1.variance(10))
-    })
-
-    it('should not collapse to 0 via cancellation for near-unit-root phi and small fractional t', () => {
-      // phi2 = 1 - 2e-14 lies just outside the 1e-14 special-case band; at
-      // t = 1e-6, 1 - phi2^t previously rounded to exactly 0 (100% error)
-      const ar1 = new AR1(Math.sqrt(1 - 2e-14), 1)
-      assert.isAbove(ar1.variance(1e-6), 0)
-    })
-  })
-
-  describe('.pdf()', () => {
-    it('should return NaN for t = 0', () => {
-      const ar1 = new AR1(0.5, 1)
-      assert(Number.isNaN(ar1.pdf(0, 0)))
-    })
-
-    it('should return NaN for t < 0', () => {
-      const ar1 = new AR1(0.5, 1)
-      assert(Number.isNaN(ar1.pdf(0, -1)))
-    })
-
-    it('should be symmetric around 0', () => {
-      const ar1 = new AR1(0.5, 1)
-      assert.closeTo(ar1.pdf(-1, 3), ar1.pdf(1, 3), 1e-10)
-    })
-  })
-
-  describe('.covariogram()', () => {
-    it('should equal variance at s = t', () => {
-      const ar1 = new AR1(0.5, 1)
-      assert.closeTo(ar1.covariogram(2, 2), ar1.variance(2), 1e-10)
-    })
-
-    it('should be symmetric: covariogram(s, t) = covariogram(t, s)', () => {
-      const ar1 = new AR1(0.5, 1)
-      assert.closeTo(ar1.covariogram(2, 3), ar1.covariogram(3, 2), 1e-10)
-    })
-
-    it('should return NaN for s < 0', () => {
-      const ar1 = new AR1(0.5, 1)
-      assert(Number.isNaN(ar1.covariogram(-1, 2)))
-    })
-
-    it('should return NaN for t < 0', () => {
-      const ar1 = new AR1(0.5, 1)
-      assert(Number.isNaN(ar1.covariogram(2, -1)))
-    })
-
-    // Cov(X_t, X_t) = Var(X_t) by definition, so this identity must hold in the same
-    // near-unit-root regime #1243 fixed in variance(); covariogram() carried the same
-    // cancellation-prone 1 - Math.pow(phi2, minTime) expression and collapsed to exactly 0.
-    it('should equal variance at s = t for near-unit-root phi and small fractional t', () => {
-      for (const [phi, sigma] of [[Math.sqrt(1 - 2e-14), 1], [Math.sqrt(1 - 5e-13), 1.5]]) {
-        const ar1 = new AR1(phi, sigma)
-        for (const t of [1e-6, 1e-4, 0.002, 0.01, 0.1]) {
-          const v = ar1.variance(t)
-          assert(v > 0, `variance(${t}) must stay positive for phi = ${phi}`)
-          assert.closeTo(ar1.covariogram(t, t) / v, 1, 1e-12,
-            `covariogram(${t}, ${t}) must match variance(${t}) for phi = ${phi}`)
-        }
-      }
-    })
-
-    // mpmath mp.dps=50, evaluating the untransformed phi^|t-s| * sigma^2 * (1-phi2^min(s,t))/(1-phi2)
-    // against the exact double phi (0x3fefffffffffffa6 = Math.sqrt(1-2e-14),
-    // 0x3feffffffffff734 = Math.sqrt(1-5e-13)) -- not the expm1 form the fix uses.
-    it('should match high-precision references for near-unit-root phi', () => {
-      const phiA = Math.sqrt(1 - 2e-14)
-      const phiB = Math.sqrt(1 - 5e-13)
-      // Each literal is the nearest double to the mpmath value shown after it.
-      const cases = [
-        // phi, sigma, s, t, reference                    mpmath mp.dps=50 value
-        [phiA, 1, 1e-6, 1e-6, 0.00000100000000000001], //  1.0000000000000099467e-6
-        [phiA, 1, 1e-6, 3e-6, 0.00000100000000000001], //  1.0000000000000099467e-6
-        [phiA, 2, 0.01, 0.03, 0.04000000000000039], //     0.040000000000000388523
-        [phiB, 1.5, 1e-4, 1e-4, 0.00022500000000005626], // 0.00022500000000005626016
-        [phiB, 1, 0.002, 0.005, 0.0020000000000004975] //  0.0020000000000004975859
-      ]
-      for (const [phi, sigma, s, t, expected] of cases) {
-        const actual = new AR1(phi, sigma).covariogram(s, t)
-        assert.closeTo(actual / expected, 1, 1e-12,
-          `covariogram(${s}, ${t}) for phi = ${phi}, sigma = ${sigma}`)
-      }
-    })
-
-    // Math.pow(phi2, 0) === 1 for every phi2 per the ECMAScript spec, so the pre-expm1
-    // formula was accidentally well-defined at min(s, t) = 0; t * Math.log(phi2) is not
-    // (0 * -Infinity = NaN when phi2 underflows to 0). X_0 = 0 deterministically, so the
-    // covariance with it is exactly 0 for every phi -- including phi = 0 and huge phi.
-    it('should return 0 when either time is 0, for every phi', () => {
-      for (const phi of [0, 1e-200, 0.5, 1, 1.5, 1e200, -0.5, -1]) {
-        const ar1 = new AR1(phi, 1)
-        assert.strictEqual(ar1.covariogram(0, 0), 0, `covariogram(0, 0) for phi = ${phi}`)
-        assert.strictEqual(ar1.covariogram(0, 3), 0, `covariogram(0, 3) for phi = ${phi}`)
-        assert.strictEqual(ar1.covariogram(3, 0), 0, `covariogram(3, 0) for phi = ${phi}`)
-      }
-    })
-
-    // The |phi2 - 1| < 1e-14 unit-root branch was the file's only uncovered line; at
-    // phi = 1 the geometric series degenerates to the random-walk covariance sigma^2*min(s,t).
-    it('should reduce to the random-walk covariance at the unit root', () => {
-      const ar1 = new AR1(1, 2)
-      // exact rational: phi = 1 => Cov(X_s, X_t) = sigma^2 * min(s, t) = 4 * 3 = 12
-      assert.closeTo(ar1.covariogram(3, 7), 12, 1e-10)
-      assert.closeTo(ar1.covariogram(7, 3), 12, 1e-10)
-    })
-  })
-
-  describe('.reset()', () => {
-    it('should restore initial state to 0', () => {
-      const ar1 = new AR1(0.5, 1)
-      ar1.seed(42)
-      for (let i = 0; i < 10; i++) ar1.next()
-      ar1.reset()
-      assert.strictEqual(ar1.state(), 0)
-    })
-  })
-
-  describe('stationarity', () => {
-    it('should converge to stationary mean 0 and variance sigma^2/(1-phi^2) across seeds', () => {
-      const phi = 0.5
-      const sigma = 1
-      // exact rational: AR1 stationary variance = sigma^2/(1-phi^2) = 1/(1-0.25) = 4/3
-      const stationaryVariance = sigma * sigma / (1 - phi * phi)
-      for (const seed of MOMENT_SEEDS) {
-        const ar1 = new AR1(phi, sigma)
-        ar1.seed(seed)
-        // burn in to reach stationarity
-        for (let i = 0; i < 500; i++) ar1.next()
-        // thin by 10: lag-10 autocorrelation = phi^10 ≈ 0.001, effectively independent
-        const samples = []
-        for (let i = 0; i < 10000; i++) {
-          ar1.next()
-          if (i % 10 === 0) samples.push(ar1.state())
-        }
-        assertSampleMoments(samples, 0, stationaryVariance, seed)
-      }
-    })
-  })
-
-  describe('explosive growth', () => {
-    it('should exhibit growing variance for |phi| > 1', () => {
-      const ar1 = new AR1(1.5, 1)
-      ar1.seed(42)
-      // run ensemble of 30 paths; at step 30 theoretical Var ≈ phi^60/1.25 ≈ 3.5e10
-      const paths = ar1.ensemble(20, 30)
-      const earlyMSV = paths.map(p => p[1] * p[1]).reduce((a, b) => a + b) / paths.length
-      const lateMSV = paths.map(p => p[30] * p[30]).reduce((a, b) => a + b) / paths.length
-      assert(lateMSV > earlyMSV * 100)
-    })
-  })
-
-  describe('.marginal()', () => {
-    it('should return a Normal instance with variance matching variance(t)', () => {
-      const ar1 = new AR1(0.5, 2)
-      const marginal = ar1.marginal(2)
-      assert.instanceOf(marginal, Normal)
-      assert.strictEqual(marginal.mean(), 0)
-      assert.closeTo(marginal.variance(), ar1.variance(2), 1e-10)
-    })
-
-    it('should have pdf matching process.pdf(x, t)', () => {
-      const ar1 = new AR1(0.5, 1)
-      const marginal = ar1.marginal(2)
-      assert.closeTo(marginal.pdf(0), ar1.pdf(0, 2), 1e-10)
-      // non-zero x exercises the z = x/s scaling term, which pdf(0) alone cannot catch
-      assert.closeTo(marginal.pdf(1), ar1.pdf(1, 2), 1e-10)
-    })
-
-    it('should invert cdf via quantile', () => {
-      const ar1 = new AR1(0.5, 1)
-      const marginal = ar1.marginal(2)
-      assert.closeTo(marginal.q(marginal.cdf(1)), 1, 1e-10)
-    })
-
-    it('should throw for t = 0', () => {
-      const ar1 = new AR1(0.5, 1)
-      assert.throws(() => ar1.marginal(0), /t must be > 0/)
-    })
-
-    it('should throw for t < 0', () => {
-      const ar1 = new AR1(0.5, 1)
-      assert.throws(() => ar1.marginal(-1), /t must be > 0/)
-    })
-
-    // A 29700-point sweep of variance(t) over phi (including a dense grid straddling the
-    // 1e-14 reformulation boundary), sigma and t found no strictly negative value for any
-    // t > 0 -- the non-stationary |phi| >= 1 branch diverges to +Infinity but never flips
-    // sign. See solutions/correctness/2026-08-01-1600-ar1-marginal-variance-guard-reachability.md
-    it('should return a positive-variance Normal on the non-stationary branch', () => {
-      for (const phi of [1, -1, 1.5, -1.5, 1 + 1e-14, 1 - 1e-14]) {
-        const ar1 = new AR1(phi, 1)
-        const marginal = ar1.marginal(3)
-        assert.instanceOf(marginal, Normal)
-        assert(ar1.variance(3) > 0, `variance must stay positive for phi = ${phi}`)
-        assert.closeTo(marginal.variance(), ar1.variance(3), 1e-10)
-      }
-    })
-
-    // v <= 0 is reachable only by floating-point underflow (t below ~1e-322, or sigma below
-    // ~1.6e-161 so that sigma^2 underflows). marginal() no longer guards this itself; the
-    // degenerate scale must still be rejected, now by Normal's own sigma > 0 validation.
-    it('should throw via Normal validation when variance underflows to zero', () => {
-      const ar1 = new AR1(0.9, 1)
-      assert.strictEqual(ar1.variance(Number.MIN_VALUE), 0)
-      assert.throws(() => ar1.marginal(Number.MIN_VALUE), /Invalid parameters/)
-
-      const tiny = new AR1(0.9, 1e-200)
-      assert.strictEqual(tiny.variance(1), 0)
-      assert.throws(() => tiny.marginal(1), /Invalid parameters/)
-    })
-  })
-
-  describe('.fit()', () => {
-    it('should recover phi and sigma from a long simulated path across seeds', () => {
-      const phi = 0.6
-      const sigma = 1.3
-      const n = 20000
-      // AR(1) OLS slope asymptotic variance (Hamilton, "Time Series Analysis", eq. 8.2.16):
-      // sqrt(n)(phiHat-phi) -> N(0, 1-phi^2), same formula OrnsteinUhlenbeck.fit()'s test uses
-      // for its OLS slope, since the true intercept being 0 here doesn't change the leading term.
-      const tolPhi = K_SIGMA * Math.sqrt((1 - phi * phi) / n)
-      // s2 = ss/(n-2) is the usual OLS residual-variance estimator; Var(s2) ~ 2*sigma^4/n
-      // (same approximation OrnsteinUhlenbeck.fit()'s test uses for its own varS2), propagated
-      // through the delta method (sigma = sqrt(s2)) for sigma_hat's tolerance.
-      const varS2 = 2 * Math.pow(sigma, 4) / n
-      const tolSigma = K_SIGMA * Math.sqrt(varS2) / (2 * sigma)
-      for (const seed of MOMENT_SEEDS) {
-        const ar1 = new AR1(phi, sigma)
-        ar1.seed(seed)
-        const fitted = AR1.fit(ar1.path(n))
-        assert.instanceOf(fitted, AR1)
-        assert.closeTo(fitted.params().phi, phi, tolPhi, `seed ${seed}: phi`)
-        assert.closeTo(fitted.params().sigma, sigma, tolSigma, `seed ${seed}: sigma`)
-      }
-    })
-
-    it('should throw when path has fewer than 4 states', () => {
-      assert.throws(() => AR1.fit([0, 1, 2]), /at least 4 states/)
-    })
-
-    it('should throw when path is not an array', () => {
-      assert.throws(() => AR1.fit(null), /at least 4 states/)
-    })
-
-    it('should succeed at exactly 4 states when the path is not perfectly collinear', () => {
-      // [0, 1, 2, 2] is not collinear (the last step breaks the straight line), so OLS
-      // residuals are not all 0 and sigma2 = ss/(n-2) is strictly positive.
-      const fitted = AR1.fit([0, 1, 2, 2])
-      assert.instanceOf(fitted, AR1)
-      assert.isTrue(Number.isFinite(fitted.params().phi))
-      assert.isAbove(fitted.params().sigma, 0)
-    })
-
-    it('should throw when a perfectly collinear path drives sigma to 0', () => {
-      // A perfectly linear path drives the OLS residuals to exactly 0, so sigma2 = 0
-      // and the AR1 constructor's Process.validate('sigma > 0') rejects it, unlike
-      // OrnsteinUhlenbeck.fit() which has its own slope-range check for the analogous case.
-      assert.throws(() => AR1.fit([0, 1, 2, 3, 4]), /sigma > 0/)
-    })
-  })
-})
-
-describe('process.Poisson', () => {
-  describe('constructor', () => {
-    it('should throw on lambda = 0', () => {
-      assert.throws(() => new ProcessPoisson(0, 1), /Invalid parameters/)
-    })
-
-    it('should throw on lambda < 0', () => {
-      assert.throws(() => new ProcessPoisson(-1, 1), /Invalid parameters/)
-    })
-
-    it('should throw on dt = 0', () => {
-      assert.throws(() => new ProcessPoisson(1, 0), /Invalid parameters/)
-    })
-
-    it('should throw on dt < 0', () => {
-      assert.throws(() => new ProcessPoisson(1, -0.5), /Invalid parameters/)
-    })
-
-    it('should throw on lambda = NaN', () => {
-      assert.throws(() => new ProcessPoisson(NaN, 1), /Invalid parameters/)
-    })
-
-    it('should accept valid parameters', () => {
-      assert.doesNotThrow(() => new ProcessPoisson(2, 0.5))
-    })
-
-    it('should start at state 0', () => {
-      const pp = new ProcessPoisson(1, 1)
-      assert.strictEqual(pp.state(), 0)
-    })
-  })
-
-  describe('path', () => {
-    it('should be non-decreasing', () => {
-      const pp = new ProcessPoisson(2, 0.1)
-      const path = pp.path(200)
-      for (let i = 1; i < path.length; i++) {
-        assert(path[i] >= path[i - 1])
-      }
-    })
-
-    it('should be integer-valued', () => {
-      const pp = new ProcessPoisson(2, 0.1)
-      const path = pp.path(200)
-      for (const x of path) {
-        assert.strictEqual(x, Math.floor(x))
-      }
-    })
-  })
-
-  describe('increments', () => {
-    it('should have mean and variance matching lambda*dt across seeds', () => {
-      const lambda = 3
-      const dt = 0.5
-      const n = 5000
-      // exact rational: Poisson(lambda*dt) increment has mean = variance = lambda*dt
-      for (const seed of MOMENT_SEEDS) {
-        const pp = new ProcessPoisson(lambda, dt)
-        pp.seed(seed)
-        assertSampleMoments(sampleSteps(pp, n), lambda * dt, lambda * dt, seed)
-      }
-    })
-  })
-
-  describe('.mean()', () => {
-    it('should return 0 at t=0', () => {
-      const pp = new ProcessPoisson(2, 0.5)
-      assert.strictEqual(pp.mean(0), 0)
-    })
-
-    it('should return NaN for t < 0', () => {
-      const pp = new ProcessPoisson(2, 0.5)
-      assert(Number.isNaN(pp.mean(-1)))
-    })
-  })
-
-  describe('.variance()', () => {
-    it('should return 0 at t=0', () => {
-      const pp = new ProcessPoisson(2, 0.5)
-      assert.strictEqual(pp.variance(0), 0)
-    })
-
-    it('should return NaN for t < 0', () => {
-      const pp = new ProcessPoisson(2, 0.5)
-      assert(Number.isNaN(pp.variance(-1)))
-    })
-  })
-
-  describe('.pdf()', () => {
-    it('should return NaN for t < 0', () => {
-      const pp = new ProcessPoisson(1, 1)
-      assert(Number.isNaN(pp.pdf(0, -1)))
-    })
-
-    it('should return 0 for non-integer x', () => {
-      const pp = new ProcessPoisson(2, 1)
-      assert.strictEqual(pp.pdf(1.5, 1), 0)
-    })
-
-    it('should return 0 for negative integer x', () => {
-      const pp = new ProcessPoisson(2, 1)
-      assert.strictEqual(pp.pdf(-1, 1), 0)
-    })
-
-    it('should return 1 for x=0 at t=0', () => {
-      const pp = new ProcessPoisson(1, 1)
-      assert.strictEqual(pp.pdf(0, 0), 1)
-    })
-
-    it('should return 0 for x=1 at t=0', () => {
-      const pp = new ProcessPoisson(1, 1)
-      assert.strictEqual(pp.pdf(1, 0), 0)
-    })
-  })
-
-  describe('.covariogram()', () => {
-    it('should be symmetric', () => {
-      const pp = new ProcessPoisson(3, 0.5)
-      assert.closeTo(pp.covariogram(2, 5), pp.covariogram(5, 2), 1e-10)
-    })
-
-    it('should equal variance at s = t', () => {
-      const pp = new ProcessPoisson(3, 0.5)
-      assert.closeTo(pp.covariogram(4, 4), pp.variance(4), 1e-10)
-    })
-
-    it('should return NaN for s < 0', () => {
-      const pp = new ProcessPoisson(1, 1)
-      assert(Number.isNaN(pp.covariogram(-1, 2)))
-    })
-
-    it('should return NaN for t < 0', () => {
-      const pp = new ProcessPoisson(1, 1)
-      assert(Number.isNaN(pp.covariogram(2, -1)))
-    })
-  })
-
-  describe('.marginal()', () => {
-    it('should return a Poisson instance with mean matching mean(t)', () => {
-      const pp = new ProcessPoisson(2, 0.5)
-      const marginal = pp.marginal(3)
-      assert.instanceOf(marginal, Poisson)
-      assert.closeTo(marginal.mean(), pp.mean(3), 1e-10)
-      assert.closeTo(marginal.variance(), pp.variance(3), 1e-10)
-    })
-
-    it('should have pdf matching process.pdf(x, t)', () => {
-      const pp = new ProcessPoisson(2, 0.5)
-      const marginal = pp.marginal(3)
-      assert.closeTo(marginal.pdf(2), pp.pdf(2, 3), 1e-10)
-    })
-
-    it('should invert cdf via quantile', () => {
-      const pp = new ProcessPoisson(2, 0.5)
-      const marginal = pp.marginal(3)
-      assert.closeTo(marginal.q(marginal.cdf(2)), 2, 1e-10)
-    })
-
-    it('should throw for t = 0', () => {
-      const pp = new ProcessPoisson(2, 0.5)
-      assert.throws(() => pp.marginal(0), /t must be > 0/)
-    })
-
-    it('should throw for t < 0', () => {
-      const pp = new ProcessPoisson(2, 0.5)
-      assert.throws(() => pp.marginal(-1), /t must be > 0/)
-    })
-  })
-
-  describe('.fit()', () => {
-    it('should recover lambda from a long simulated path across seeds', () => {
-      const lambda = 3
-      const dt = 0.5
-      const n = 20000
-      // fit() is the exact MLE: the total count over n*dt observed time is itself exactly
-      // Poisson(n*dt*lambda) (sum of n i.i.d. Poisson(lambda*dt) increments), so
-      // Var(lambdaHat) = Var(totalCount)/(n*dt)^2 = lambda/(n*dt).
-      const tol = K_SIGMA * Math.sqrt(lambda / (n * dt))
-      for (const seed of MOMENT_SEEDS) {
-        const pp = new ProcessPoisson(lambda, dt)
-        pp.seed(seed)
-        const fitted = ProcessPoisson.fit(pp.path(n), dt)
-        assert.instanceOf(fitted, ProcessPoisson)
-        assert.closeTo(fitted.params().lambda, lambda, tol, `seed ${seed}: lambda`)
-      }
-    })
-
-    it('should default dt to 1', () => {
-      const pp = new ProcessPoisson(2, 1)
-      pp.seed(1)
-      const fitted = ProcessPoisson.fit(pp.path(20000))
-      assert.strictEqual(fitted.params().dt, 1)
-    })
-
-    it('should throw when path has fewer than 2 states', () => {
-      assert.throws(() => ProcessPoisson.fit([0], 1), /at least 2 states/)
-    })
-
-    it('should throw on dt = 0', () => {
-      assert.throws(() => ProcessPoisson.fit([0, 1, 2], 0), /Invalid parameters/)
-    })
-
-    it('should throw on dt < 0', () => {
-      assert.throws(() => ProcessPoisson.fit([0, 1, 2], -1), /Invalid parameters/)
-    })
-
-    it('should throw when the path decreases', () => {
-      assert.throws(() => ProcessPoisson.fit([0, 1, 2, 1], 1), /non-decreasing/)
-    })
-
-    it('should throw when no arrivals are observed', () => {
-      assert.throws(() => ProcessPoisson.fit([0, 0, 0], 1), /lambda is not positive/)
-    })
-  })
-})
-
-describe('process.PoissonProcess (deprecated alias)', () => {
-  it('should be an instance of Poisson and behave identically', () => {
-    const pp = withSuppressedWarnings(() => new PoissonProcess(2, 0.5))
-    assert(pp instanceof ProcessPoisson)
-    assert.strictEqual(pp.mean(3), new ProcessPoisson(2, 0.5).mean(3))
-    assert.strictEqual(pp.pdf(1, 3), new ProcessPoisson(2, 0.5).pdf(1, 3))
-  })
-
-  it('should emit a deprecation warning naming the replacement', () => {
-    let pp
-    const warnings = captureWarnings(() => { pp = new PoissonProcess(2, 0.5) })
-    assert(pp instanceof ProcessPoisson)
-    assert.strictEqual(warnings.length, 1)
-    assert.match(warnings[0], /ran\.process\.PoissonProcess is deprecated and will be removed in v1\.33\.0; use ran\.process\.Poisson instead\./)
-  })
-
-  it('should have .fit() return a PoissonProcess instance, not a plain Poisson', () => {
-    const fitted = withSuppressedWarnings(() => {
-      const pp = new PoissonProcess(2, 1)
-      pp.seed(1)
-      return PoissonProcess.fit(pp.path(5000))
-    })
-    assert.instanceOf(fitted, PoissonProcess)
-  })
-})
-
-describe('process.CompoundPoisson', () => {
-  describe('constructor', () => {
-    it('should throw on lambda = 0', () => {
-      assert.throws(() => new CompoundPoisson(new Normal(1, 1), 0, 1), /Invalid parameters/)
-    })
-
-    it('should throw on lambda < 0', () => {
-      assert.throws(() => new CompoundPoisson(new Normal(1, 1), -1, 1), /Invalid parameters/)
-    })
-
-    it('should throw on dt = 0', () => {
-      assert.throws(() => new CompoundPoisson(new Normal(1, 1), 1, 0), /Invalid parameters/)
-    })
-
-    it('should throw on dt < 0', () => {
-      assert.throws(() => new CompoundPoisson(new Normal(1, 1), 1, -0.5), /Invalid parameters/)
-    })
-
-    it('should throw on lambda = NaN', () => {
-      assert.throws(() => new CompoundPoisson(new Normal(1, 1), NaN, 1), /Invalid parameters/)
-    })
-
-    it('should throw when jumpDist is undefined', () => {
-      assert.throws(() => new CompoundPoisson(undefined, 1, 1), /Invalid parameters/)
-    })
-
-    it('should throw when jumpDist is null', () => {
-      assert.throws(() => new CompoundPoisson(null, 1, 1), /Invalid parameters/)
-    })
-
-    it('should throw when jumpDist has no .sample() method', () => {
-      assert.throws(() => new CompoundPoisson({ mean: () => 0 }, 1, 1), /Invalid parameters/)
-    })
-
-    it('should accept valid parameters', () => {
-      assert.doesNotThrow(() => new CompoundPoisson(new Normal(1, 1), 2, 0.5))
-    })
-
-    it('should accept default lambda and dt when jumpDist is provided', () => {
-      assert.doesNotThrow(() => new CompoundPoisson(new Normal(0, 1), 1, 1))
-    })
-
-    it('should start at state 0', () => {
-      const cpp = new CompoundPoisson(new Normal(1, 1), 2, 1)
-      assert.strictEqual(cpp.state(), 0)
-    })
-  })
-
-  describe('path', () => {
-    it('should start at 0', () => {
-      const cpp = new CompoundPoisson(new Poisson(1), 2, 1)
-      assert.strictEqual(cpp.path(10)[0], 0)
-    })
-
-    it('should be non-decreasing with non-negative jumps', () => {
-      const cpp = new CompoundPoisson(new Poisson(2), 3, 0.5)
-      cpp.seed(42)
-      const path = cpp.path(200)
-      for (let i = 1; i < path.length; i++) {
-        assert(path[i] >= path[i - 1])
-      }
-    })
-  })
-
-  describe('.mean()', () => {
-    it('should return 0 at t=0', () => {
-      const cpp = new CompoundPoisson(new Normal(2, 1), 3, 1)
-      assert.strictEqual(cpp.mean(0), 0)
-    })
-
-    it('should return NaN for t < 0', () => {
-      const cpp = new CompoundPoisson(new Normal(1, 1), 2, 1)
-      assert(Number.isNaN(cpp.mean(-1)))
-    })
-  })
-
-  describe('.variance()', () => {
-    it('should return 0 at t=0', () => {
-      const cpp = new CompoundPoisson(new Normal(1, 1), 2, 1)
-      assert.strictEqual(cpp.variance(0), 0)
-    })
-
-    it('should return NaN for t < 0', () => {
-      const cpp = new CompoundPoisson(new Normal(1, 1), 2, 1)
-      assert(Number.isNaN(cpp.variance(-1)))
-    })
-  })
-
-  describe('.covariogram()', () => {
-    it('should be symmetric', () => {
-      const cpp = new CompoundPoisson(new Normal(2, 1), 3, 1)
-      assert.closeTo(cpp.covariogram(2, 5), cpp.covariogram(5, 2), 1e-10)
-    })
-
-    it('should equal variance at s = t', () => {
-      const cpp = new CompoundPoisson(new Normal(2, 1), 3, 1)
-      // exact rational: variance(3) = covariogram(3, 3) = 3*3*(1+4) = 45
-      assert.closeTo(cpp.covariogram(3, 3), cpp.variance(3), 1e-10)
-    })
-
-    it('should return NaN for s < 0', () => {
-      const cpp = new CompoundPoisson(new Normal(1, 1), 2, 1)
-      assert(Number.isNaN(cpp.covariogram(-1, 2)))
-    })
-
-    it('should return NaN for t < 0', () => {
-      const cpp = new CompoundPoisson(new Normal(1, 1), 2, 1)
-      assert(Number.isNaN(cpp.covariogram(2, -1)))
-    })
-  })
-
-  describe('.seed()', () => {
-    it('should produce identical paths when seeded identically', () => {
-      const cpp = new CompoundPoisson(new Normal(1, 1), 2, 1)
-      cpp.seed(42)
-      const path1 = cpp.path(30)
-      cpp.seed(42)
-      const path2 = cpp.path(30)
-      assert.deepEqual(path1, path2)
-    })
-
-    it('should produce different paths for different seeds', () => {
-      const cpp = new CompoundPoisson(new Normal(1, 1), 2, 1)
-      cpp.seed(1)
-      const path1 = cpp.path(30)
-      cpp.seed(2)
-      const path2 = cpp.path(30)
-      assert.notDeepEqual(path1, path2)
-    })
-
-    it('should return this for chaining', () => {
-      const cpp = new CompoundPoisson(new Normal(1, 1), 2, 1)
-      assert.strictEqual(cpp.seed(0), cpp)
-    })
-  })
-
-  describe('.params()', () => {
-    it('returns a jumpDist that is not the same reference as the live instance\'s', () => {
-      const cpp = new CompoundPoisson(new Gamma(2, 1), 3, 1)
-      assert.notStrictEqual(cpp.params().jumpDist, cpp.p.jumpDist)
-    })
-
-    it('seeding params().jumpDist does not desync the live process\'s future path', () => {
-      // Reproduces the aliasing bug: _next() samples from this.p.jumpDist directly every step
-      // with no per-step reseed (unlike CompoundPoisson.prototype.seed(), which only reseeds
-      // jumpDist once), so a live reference let an external .seed() call on the "snapshot"
-      // silently desync the process's own reproducible stream.
-      const cpp1 = new CompoundPoisson(new Gamma(2, 1), 3, 1).seed(42)
-      const path1 = cpp1.path(20)
-
-      const cpp2 = new CompoundPoisson(new Gamma(2, 1), 3, 1).seed(42)
-      cpp2.params().jumpDist.seed(999)
-      const path2 = cpp2.path(20)
-
-      assert.deepEqual(path1, path2)
-    })
-  })
-
-  describe('.reset()', () => {
-    it('should restore initial state to 0', () => {
-      const cpp = new CompoundPoisson(new Normal(0, 1), 2, 1)
-      cpp.seed(42)
-      for (let i = 0; i < 10; i++) cpp.next()
-      cpp.reset()
-      assert.strictEqual(cpp.state(), 0)
-    })
-  })
-
-  describe('increments', () => {
-    it('should have mean and variance matching lambda*dt*E[J] and lambda*dt*E[J^2] across seeds', () => {
-      const lambda = 2
-      const dt = 0.5
-      const muJ = 1
-      const n = 5000
-      // exact rational: compound Poisson increment has mean = lambda*dt*E[J], variance = lambda*dt*E[J^2];
-      // for jumpDist = Normal(muJ, 1), E[J] = muJ and E[J^2] = Var(J) + E[J]^2 = 1 + muJ^2
-      const expectedMean = lambda * dt * muJ
-      const expectedVariance = lambda * dt * (1 + muJ * muJ)
-      for (const seed of MOMENT_SEEDS) {
-        const cpp = new CompoundPoisson(new Normal(muJ, 1), lambda, dt)
-        cpp.seed(seed)
-        assertSampleMoments(sampleResetSteps(cpp, n), expectedMean, expectedVariance, seed)
-      }
-    })
-  })
-
-  describe('.marginal()', () => {
-    it('should throw a CompoundPoisson-specific error explaining the general case is unsupported', () => {
-      const cpp = new CompoundPoisson(new Normal(1, 1), 2, 1)
-      assert.throws(() => cpp.marginal(1), /CompoundPoisson\.marginal\(\): no closed form exists for an arbitrary jump distribution/)
-    })
-
-    it('should throw for t <= 0 even with a Gamma jumpDist', () => {
-      const cpp = new CompoundPoisson(new Gamma(2, 3), 2, 1)
-      assert.throws(() => cpp.marginal(0), /t must be > 0/)
-      assert.throws(() => cpp.marginal(-1), /t must be > 0/)
-    })
-
-    it('should return a Tweedie instance for a Gamma jumpDist, matching mean and variance', () => {
-      const alpha = 2
-      const beta = 3
-      const lambda = 2
-      const t = 1.5
-      const cpp = new CompoundPoisson(new Gamma(alpha, beta), lambda, 1)
-      const marginal = cpp.marginal(t)
-      assert.instanceOf(marginal, Tweedie)
-      assert.closeTo(marginal.mean(), cpp.mean(t), 1e-10)
-      assert.closeTo(marginal.variance(), cpp.variance(t), 1e-10)
-    })
-
-    it('should return a Tweedie marginal whose pdf/cdf match the exact Poisson-gamma mixture', () => {
-      // exact rational: P(Y=0) = P(N=0) = exp(-lambda*t) for the compound Poisson-gamma mixture,
-      // which is also Tweedie's own point-mass formula -- an independent closed-form check that
-      // does not rely on the marginal()/Tweedie parameter-mapping algebra under test.
-      const alpha = 2
-      const beta = 3
-      const lambda = 2
-      const t = 1.5
-      const cpp = new CompoundPoisson(new Gamma(alpha, beta), lambda, 1)
-      const marginal = cpp.marginal(t)
-      assert.closeTo(marginal.pdf(0), Math.exp(-lambda * t), 1e-10)
-      assert.closeTo(marginal.cdf(0), Math.exp(-lambda * t), 1e-10)
-
-      // Independent reference for an interior point (x = mu, the Tweedie mean), built directly
-      // from the compound Poisson-gamma series rather than the marginal()/Tweedie parameter
-      // mapping under test: X_t = sum_{i=1}^{N} J_i with N ~ Poisson(lambda*t) and
-      // J_i ~ Gamma(alpha, beta), so J_1 + ... + J_n ~ Gamma(n*alpha, beta) for n >= 1 (sum of
-      // n i.i.d. Gamma(alpha, beta) variates), giving
-      // pdf(x) = sum_{n=1}^{inf} Poisson(lambda*t).pdf(n) * Gamma(n*alpha, beta).pdf(x) and
-      // cdf(x) = exp(-lambda*t)*(x>=0) + sum_{n=1}^{inf} Poisson(lambda*t).pdf(n) * Gamma(n*alpha, beta).cdf(x).
-      // The series is truncated at N = 30, where the Poisson(3) tail 1 - cdf(30) is far below
-      // 1e-12, so truncation error is negligible next to the comparison tolerance.
-      // mu = lambda*t*alpha/beta = 2*1.5*2/3 (exact rational) = 2.
-      const mu = 2
-      // mpmath mp.dps=50: sum_{n=1}^{30} Poisson(3).pmf(n) * Gamma(2n, scale=1/3).pdf(2), with
-      // Gamma.pdf/cdf and Poisson.pmf reimplemented from scratch via mpmath's own
-      // exp/log/loggamma/gammainc -> 0.26858960310934577112955117114307537913148585514259
-      const pdfRef = 0.26858960310934576
-      // mpmath mp.dps=50: exp(-3) + sum_{n=1}^{30} Poisson(3).pmf(n) * Gamma(2n, scale=1/3).cdf(2)
-      // -> 0.56348946343750803140234169015766437926068330168143
-      const cdfRef = 0.563489463437508
-      assert.closeTo(marginal.pdf(mu), pdfRef, 1e-8)
-      assert.closeTo(marginal.cdf(mu), cdfRef, 1e-8)
-    })
-  })
-
-  describe('.fit()', () => {
-    it('should recover lambda and the jump distribution from a long simulated path across seeds', () => {
-      const muJ = 2
-      const sigmaJ = 0.5
-      const lambda = 0.5
-      const dt = 0.02
-      const n = 200000
-      // fit() treats every non-zero increment as exactly one jump, since individual arrival
-      // counts within a dt interval are not observable from the cumulative path alone. This is
-      // exact only when at most one jump ever lands in the same interval; at lambda*dt = 0.01
-      // the probability of a merged 2+-jump interval is negligible (~5e-5), so the systematic
-      // bias this approximation introduces is far below the sampling noise below.
-      const p = 1 - Math.exp(-lambda * dt)
-      const m = n * p // expected number of detected jumps
-      const tolLambda = K_SIGMA * Math.sqrt(lambda / (n * dt))
-      const tolMuJ = K_SIGMA * sigmaJ / Math.sqrt(m)
-      const tolSigmaJ = K_SIGMA * sigmaJ / Math.sqrt(2 * (m - 1))
-      for (const seed of MOMENT_SEEDS) {
-        const cpp = new CompoundPoisson(new Normal(muJ, sigmaJ), lambda, dt)
-        cpp.seed(seed)
-        const fitted = CompoundPoisson.fit(cpp.path(n), dt, Normal)
-        assert.instanceOf(fitted, CompoundPoisson)
-        assert.closeTo(fitted.params().lambda, lambda, tolLambda, `seed ${seed}: lambda`)
-        assert.instanceOf(fitted.params().jumpDist, Normal)
-        const { mu: fittedMuJ, sigma: fittedSigmaJ } = fitted.params().jumpDist.params()
-        assert.closeTo(fittedMuJ, muJ, tolMuJ, `seed ${seed}: muJ`)
-        assert.closeTo(fittedSigmaJ, sigmaJ, tolSigmaJ, `seed ${seed}: sigmaJ`)
-      }
-    })
-
-    it('should default dt to 1', () => {
-      const cpp = new CompoundPoisson(new Normal(0, 1), 2, 1)
-      cpp.seed(1)
-      const fitted = CompoundPoisson.fit(cpp.path(20000), undefined, Normal)
-      assert.strictEqual(fitted.params().dt, 1)
-    })
-
-    it('should throw when path has fewer than 2 states', () => {
-      assert.throws(() => CompoundPoisson.fit([0], 1, Normal), /at least 2 states/)
-    })
-
-    it('should throw on dt = 0', () => {
-      assert.throws(() => CompoundPoisson.fit([0, 1, 2], 0, Normal), /Invalid parameters/)
-    })
-
-    it('should throw on dt < 0', () => {
-      assert.throws(() => CompoundPoisson.fit([0, 1, 2], -1, Normal), /Invalid parameters/)
-    })
-
-    it('should throw when jumpDistConstructor has no static fit()', () => {
-      assert.throws(() => CompoundPoisson.fit([0, 1, 2], 1, {}), /static fit\(\) method/)
-    })
-
-    it('should throw when jumpDistConstructor is missing', () => {
-      assert.throws(() => CompoundPoisson.fit([0, 1, 2], 1), /static fit\(\) method/)
-    })
-
-    it('should throw when the path has no non-zero increments', () => {
-      assert.throws(() => CompoundPoisson.fit([0, 0, 0], 1, Normal), /no non-zero increments/)
-    })
-  })
-})
-
-describe('process.CompoundPoissonProcess (deprecated alias)', () => {
-  it('should be an instance of CompoundPoisson and behave identically', () => {
-    const cpp = withSuppressedWarnings(() => new CompoundPoissonProcess(new Normal(1, 1), 2, 1))
-    assert(cpp instanceof CompoundPoisson)
-    assert.strictEqual(cpp.mean(3), new CompoundPoisson(new Normal(1, 1), 2, 1).mean(3))
-  })
-
-  it('should emit a deprecation warning naming the replacement', () => {
-    let cpp
-    const warnings = captureWarnings(() => { cpp = new CompoundPoissonProcess(new Normal(1, 1), 2, 1) })
-    assert(cpp instanceof CompoundPoisson)
-    assert.strictEqual(warnings.length, 1)
-    assert.match(warnings[0], /ran\.process\.CompoundPoissonProcess is deprecated and will be removed in v1\.33\.0; use ran\.process\.CompoundPoisson instead\./)
-  })
-
-  it('should have .fit() return a CompoundPoissonProcess instance, not a plain CompoundPoisson', () => {
-    const fitted = withSuppressedWarnings(() => {
-      const cpp = new CompoundPoissonProcess(new Normal(2, 0.5), 2, 1)
-      cpp.seed(1)
-      return CompoundPoissonProcess.fit(cpp.path(5000), 1, Normal)
-    })
-    assert.instanceOf(fitted, CompoundPoissonProcess)
-  })
-})
-
-describe('process.CoxIngersollRoss', () => {
-  describe('constructor', () => {
-    it('should throw on kappa = 0', () => {
-      assert.throws(() => new CoxIngersollRoss(0, 1, 1, 1), /Invalid parameters/)
-    })
-
-    it('should throw on kappa < 0', () => {
-      assert.throws(() => new CoxIngersollRoss(-1, 1, 1, 1), /Invalid parameters/)
-    })
-
-    it('should throw on theta = 0', () => {
-      assert.throws(() => new CoxIngersollRoss(1, 0, 1, 1), /Invalid parameters/)
-    })
-
-    it('should throw on theta < 0', () => {
-      assert.throws(() => new CoxIngersollRoss(1, -1, 1, 1), /Invalid parameters/)
-    })
-
-    it('should throw on sigma = 0', () => {
-      assert.throws(() => new CoxIngersollRoss(1, 1, 0, 1), /Invalid parameters/)
-    })
-
-    it('should throw on sigma < 0', () => {
-      assert.throws(() => new CoxIngersollRoss(1, 1, -1, 1), /Invalid parameters/)
-    })
-
-    it('should throw on dt = 0', () => {
-      assert.throws(() => new CoxIngersollRoss(1, 1, 1, 0), /Invalid parameters/)
-    })
-
-    it('should throw on dt < 0', () => {
-      assert.throws(() => new CoxIngersollRoss(1, 1, 1, -0.5), /Invalid parameters/)
-    })
-
-    it('should throw on kappa = NaN', () => {
-      assert.throws(() => new CoxIngersollRoss(NaN, 1, 1, 1), /Invalid parameters/)
-    })
-
-    it('should accept valid parameters', () => {
-      assert.doesNotThrow(() => new CoxIngersollRoss(2, 1, 0.5, 0.1))
-    })
-
-    it('should not throw when Feller condition is not met', () => {
-      // 2*0.5*1 = 1 <= 4 = sigma^2; Feller not satisfied, but only a warning
-      assert.doesNotThrow(() => new CoxIngersollRoss(0.5, 1, 2, 1))
-    })
-
-    it('should start at state 0', () => {
-      const cir = new CoxIngersollRoss(2, 1, 0.5, 0.1)
-      assert.strictEqual(cir.state(), 0)
-    })
-  })
-
-  describe('.mean()', () => {
-    it('should return 0 at t=0', () => {
-      const cir = new CoxIngersollRoss(2, 3, 1, 0.1)
-      assert.strictEqual(cir.mean(0), 0)
-    })
-
-    it('should return NaN for t < 0', () => {
-      const cir = new CoxIngersollRoss(1, 1, 1, 1)
-      assert(isNaN(cir.mean(-1)))
-    })
-
-    it('should be stable after advancing the simulation', () => {
-      const cir = new CoxIngersollRoss(2, 3, 1, 0.1)
-      const before = cir.mean(1)
-      for (let i = 0; i < 20; i++) cir.next()
-      assert.closeTo(cir.mean(1), before, 1e-10)
-    })
-  })
-
-  describe('.variance()', () => {
-    it('should return 0 at t=0', () => {
-      const cir = new CoxIngersollRoss(1, 1, 1, 1)
-      assert.strictEqual(cir.variance(0), 0)
-    })
-
-    it('should return NaN for t < 0', () => {
-      const cir = new CoxIngersollRoss(1, 1, 1, 1)
-      assert(isNaN(cir.variance(-1)))
-    })
-  })
-
-  describe('positivity', () => {
-    it('should produce non-negative paths when Feller condition holds', () => {
-      // 2*kappa*theta = 2*2*1 = 4 > 1 = sigma^2; Feller condition met
-      const cir = new CoxIngersollRoss(2, 1, 1, 0.01)
-      cir.seed(0)
-      const path = cir.path(10000)
-      assert(path.every(x => x >= 0))
-    })
-  })
-
-  describe('stationarity', () => {
-    it('should converge to stationary mean theta and variance sigma^2*theta/(2*kappa) across seeds', () => {
-      const kappa = 2; const theta = 1.5; const sigma = 0.5; const dt = 0.01
-      const burnInSteps = 500
-      const m = 3000
-      // exact rational: CIR stationary distribution is Gamma with mean theta and variance
-      // sigma^2*theta/(2*kappa); burn-in time = burnInSteps*dt = 5 with kappa=2 makes the
-      // finite-time bias (~exp(-kappa*burnInTime) = exp(-10)) negligible relative to the
-      // sampling tolerance
-      const stationaryVariance = sigma * sigma * theta / (2 * kappa)
-      for (const seed of MOMENT_SEEDS) {
-        const cir = new CoxIngersollRoss(kappa, theta, sigma, dt)
-        cir.seed(seed)
-        const paths = cir.ensemble(m, burnInSteps)
-        const samples = paths.map(path => path[burnInSteps])
-        assertSampleMoments(samples, theta, stationaryVariance, seed)
-      }
-    })
-  })
-
-  describe('.reset()', () => {
-    it('should restore initial state to 0', () => {
-      const cir = new CoxIngersollRoss(2, 1, 0.5, 0.1)
-      for (let i = 0; i < 10; i++) cir.next()
-      cir.reset()
-      assert.strictEqual(cir.state(), 0)
-    })
-  })
-
-  describe('.pdf()', () => {
-    it('should return NaN for t = 0', () => {
-      const cir = new CoxIngersollRoss(2, 3, 1, 0.1)
-      assert(Number.isNaN(cir.pdf(1, 0)))
-    })
-
-    it('should return NaN for t < 0', () => {
-      const cir = new CoxIngersollRoss(2, 3, 1, 0.1)
-      assert(Number.isNaN(cir.pdf(1, -1)))
-    })
-
-    it('should return 0 for x < 0', () => {
-      const cir = new CoxIngersollRoss(2, 3, 1, 0.1)
-      assert.strictEqual(cir.pdf(-0.1, 0.5), 0)
-    })
-
-    it('should return 0 for x = 0 when Feller condition holds (alpha > 1)', () => {
-      // kappa=2, theta=3, sigma=1: alpha = 2*kappa*theta/sigma^2 = 12 > 1
-      const cir = new CoxIngersollRoss(2, 3, 1, 0.1)
-      assert.strictEqual(cir.pdf(0, 0.5), 0)
-    })
-
-    it('should be stable after advancing the simulation', () => {
-      const cir = new CoxIngersollRoss(2, 3, 1, 0.1)
-      const before = cir.pdf(1.0, 0.5)
-      for (let i = 0; i < 20; i++) cir.next()
-      assert.closeTo(cir.pdf(1.0, 0.5), before, 1e-10)
-    })
-  })
-
-  describe('.covariogram()', () => {
-    it('should return NaN for s < 0', () => {
-      const cir = new CoxIngersollRoss(2, 3, 1, 0.1)
-      assert(Number.isNaN(cir.covariogram(-1, 2)))
-    })
-
-    it('should return NaN for t < 0', () => {
-      const cir = new CoxIngersollRoss(2, 3, 1, 0.1)
-      assert(Number.isNaN(cir.covariogram(2, -1)))
-    })
-
-    it('should be symmetric', () => {
-      const cir = new CoxIngersollRoss(2, 3, 1, 0.1)
-      assert.closeTo(cir.covariogram(1, 3), cir.covariogram(3, 1), 1e-10)
-    })
-
-    it('should equal variance at s = t', () => {
-      const cir = new CoxIngersollRoss(2, 3, 1, 0.1)
-      // mpmath mp.dps=50: theta*sigma^2/(2*kappa)*(1-exp(-kappa*t))^2 at t=2 ->
-      // 0.72277813863782561343853900993447378244661507377008. Documentation only — the assertion
-      // below is the structural cov(t,t) = variance(t) identity, not a reference-value check.
-      assert.closeTo(cir.covariogram(2, 2), cir.variance(2), 1e-10)
-    })
-
-    it('should return 0 at s=0 or t=0', () => {
-      const cir = new CoxIngersollRoss(2, 3, 1, 0.1)
-      // exact rational: min(s,t)=0 so (1-exp(0))^2 = 0
-      assert.strictEqual(cir.covariogram(0, 2), 0)
-      assert.strictEqual(cir.covariogram(2, 0), 0)
-    })
-  })
-
-  describe('.marginal()', () => {
-    it('should return a Gamma distribution matching pdf() at a given point', () => {
-      const cir = new CoxIngersollRoss(2, 3, 1, 0.1)
-      const marginal = cir.marginal(0.5)
-      assert.instanceOf(marginal, Gamma)
-      assert.closeTo(marginal.pdf(0.5), cir.pdf(0.5, 0.5), 1e-10)
-      assert.closeTo(marginal.pdf(2.0), cir.pdf(2.0, 0.5), 1e-10)
-    })
-
-    it('should have mean and variance matching mean()/variance()', () => {
-      const cir = new CoxIngersollRoss(2, 3, 1, 0.1)
-      const marginal = cir.marginal(1)
-      assert.closeTo(marginal.mean(), cir.mean(1), 1e-10)
-      assert.closeTo(marginal.variance(), cir.variance(1), 1e-10)
-    })
-
-    it('should match pdf() at x=0 when alpha <= 1 (Feller condition violated)', () => {
-      // alpha = 2*kappa*theta/sigma^2 = 2*1*0.4/4 = 0.2 < 1
-      const cir = new CoxIngersollRoss(1, 0.4, 2, 0.1)
-      const marginal = cir.marginal(0.5)
-      assert.strictEqual(marginal.pdf(0), cir.pdf(0, 0.5))
-    })
-
-    it('should round-trip quantile(cdf(x)) = x, exercising the Distribution API beyond pdf/mean/variance', () => {
-      const cir = new CoxIngersollRoss(2, 3, 1, 0.1)
-      const marginal = cir.marginal(1)
-      assert.closeTo(marginal.q(marginal.cdf(2)), 2, 1e-10)
-    })
-
-    it('should throw for t = 0', () => {
-      const cir = new CoxIngersollRoss(2, 3, 1, 0.1)
-      assert.throws(() => cir.marginal(0), /t must be > 0/)
-    })
-
-    it('should throw for t < 0', () => {
-      const cir = new CoxIngersollRoss(2, 3, 1, 0.1)
-      assert.throws(() => cir.marginal(-1), /t must be > 0/)
-    })
-  })
-
-  describe('.fit()', () => {
-    it('should recover kappa, theta, and sigma from a long simulated path across seeds', () => {
-      const kappa = 2
-      const theta = 3
-      const sigma = 1
-      const dt = 0.01
-      const n = 100000
-      // Stage 1 (kappa, theta) regresses X_{n+1} on X_n exactly like OrnsteinUhlenbeck.fit() —
-      // same b = exp(-kappa*dt), a = theta*(1-b) back-substitution — so Hamilton's classical
-      // AR(1)-OLS asymptotic slope variance (sqrt(n)(bHat-b) -> N(0, 1-b^2)) applies to kappa via
-      // the same delta method OU uses for its theta. Unlike OU, CIR's one-step conditional
-      // variance is heteroskedastic (proportional to X_n, from the sigma*sqrt(X) diffusion term)
-      // rather than OU's flat s2; evaluating the exact conditional variance (the same
-      // Var[X_{n+1}|X_n] = X_n*sigma^2/kappa*(b-b^2) + theta*sigma^2/(2*kappa)*(1-b)^2 that
-      // fit()'s stage 2 estimates) at the stationary mean X_n = theta gives the representative
-      // constant-variance analogue of OU's s2, used below for theta's tolerance.
-      const b = Math.exp(-kappa * dt)
-      const s2 = theta * sigma * sigma / (2 * kappa) * (1 - b * b)
-      const varB = (1 - b * b) / n
-      // kappa = -ln(b)/dt, d(kappa)/d(b) = -1/(b*dt) — identical delta method to OU's theta.
-      const varKappa = varB / (b * b * dt * dt)
-      // theta = a/(1-b) — same OLS delta-method form as OU's varMu, with CIR's heteroskedastic
-      // s2 (above) standing in for OU's constant one.
-      const varTheta = s2 / (n * (1 - b) * (1 - b))
-      // sigma^2 = alpha*kappa/(b-b^2), where alpha is stage 2's OLS slope of e_i^2 on X_i. Unlike
-      // OU's s2Hat (a plain sample mean of homoskedastic squared Gaussian residuals), alpha_hat
-      // is itself a regression slope on heteroskedastic, non-Gaussian e_i^2, so its sampling
-      // variance needs the OLS-slope sandwich form, not a chi-squared sample-variance one. dt is
-      // small here (kappa*dt = 0.02), so the noncentral chi-squared driving each step has
-      // noncentrality far exceeding its degrees of freedom (~1000x at X = theta) and is close to
-      // Gaussian by the standard noncentral-chi-squared-to-normal limit; under that
-      // approximation e_i^2's residual variance around its mean s2 is ~2*s2^2 (a Gaussian
-      // fourth-moment), giving the same sandwich form used for varB: Var(alpha_hat) ~
-      // 2*s2^2/(n*Var(X)), Var(X) = the stationary variance theta*sigma^2/(2*kappa). We
-      // conservatively double this estimate: e_i^2's true distribution (noncentral-chi-squared-
-      // driven, not exactly Gaussian) has excess kurtosis this approximation doesn't capture, and
-      // CLS is documented to be less efficient than exact MLE on this second-moment stage.
-      const vStat = theta * sigma * sigma / (2 * kappa)
-      const varAlpha = 2 * (2 * s2 * s2 / (n * vStat))
-      const alpha = sigma * sigma / kappa * (b - b * b)
-      const dSigma2dAlpha = kappa / (b - b * b)
-      // Stage 1's kappa/b uncertainty also feeds into sigma^2 = alpha*kappa/(b-b^2); propagated
-      // the same way as OU's dSigma2dB (kappa = -ln(b)/dt substituted in before differentiating).
-      const dSigma2dB = alpha * ((b - 1) + Math.log(b) * (1 - 2 * b)) / (dt * (b - b * b) * (b - b * b))
-      const varSigma2 = dSigma2dAlpha * dSigma2dAlpha * varAlpha + dSigma2dB * dSigma2dB * varB
-      const varSigma = varSigma2 / (4 * sigma * sigma)
-      // These three MOMENT_SEEDS are fixed and deterministic — unlike a test that resamples a
-      // fresh seed every run, they never change, so K_SIGMA=8's enormous margin (guarding against
-      // flakiness on arbitrary future seeds) buys nothing extra here. K=5 still corresponds to a
-      // ~5.7e-7 two-sided tail probability under the CLT normal approximation above — vastly more
-      // than three already-known, fixed outcomes need — while giving a materially tighter band
-      // than the old flat 20% (verified empirically to hold with comfortable margin, including
-      // against a 300-seed sweep well beyond the 3 seeds actually asserted below).
-      const K = 5
-      const tolKappa = K * Math.sqrt(varKappa)
-      const tolTheta = K * Math.sqrt(varTheta)
-      const tolSigma = K * Math.sqrt(varSigma)
-      for (const seed of MOMENT_SEEDS) {
-        const cir = new CoxIngersollRoss(kappa, theta, sigma, dt)
-        cir.seed(seed)
-        const fitted = CoxIngersollRoss.fit(cir.path(n), dt)
-        assert.instanceOf(fitted, CoxIngersollRoss)
-        assert.closeTo(fitted.params().kappa, kappa, tolKappa, `seed ${seed}: kappa`)
-        assert.closeTo(fitted.params().theta, theta, tolTheta, `seed ${seed}: theta`)
-        assert.closeTo(fitted.params().sigma, sigma, tolSigma, `seed ${seed}: sigma`)
-      }
-    })
-
-    it('should default dt to 1', () => {
-      const cir = new CoxIngersollRoss(2, 3, 1, 0.01)
-      cir.seed(1)
-      const fitted = CoxIngersollRoss.fit(cir.path(20000))
-      assert.strictEqual(fitted.params().dt, 1)
-    })
-
-    it('should throw when path has fewer than 4 states', () => {
-      assert.throws(() => CoxIngersollRoss.fit([1, 2, 3], 1), /at least 4 states/)
-    })
-
-    it('should throw when the AR(1) slope estimate is out of (0,1)', () => {
-      assert.throws(() => CoxIngersollRoss.fit([1, 5, 2, 5, 3, 5], 1), /AR\(1\) slope/)
-    })
-
-    it('should throw when the estimated sigma^2 is non-positive', () => {
-      assert.throws(() => CoxIngersollRoss.fit([5, 4, 3, 2, 1, 2], 1), /sigma\^2 is non-positive/)
-    })
-  })
-})
-
-describe('process.RandomWalk', () => {
-  describe('constructor', () => {
-    it('should throw on p = 0', () => {
-      assert.throws(() => new RandomWalk(0), /Invalid parameters/)
-    })
-
-    it('should throw on p = 1', () => {
-      assert.throws(() => new RandomWalk(1), /Invalid parameters/)
-    })
-
-    it('should throw on p < 0', () => {
-      assert.throws(() => new RandomWalk(-0.1), /Invalid parameters/)
-    })
-
-    it('should throw on p > 1', () => {
-      assert.throws(() => new RandomWalk(1.1), /Invalid parameters/)
-    })
-
-    it('should throw on p = NaN', () => {
-      assert.throws(() => new RandomWalk(NaN), /Invalid parameters/)
-    })
-
-    it('should accept valid probability', () => {
-      assert.doesNotThrow(() => new RandomWalk(0.3))
-      assert.doesNotThrow(() => new RandomWalk(0.7))
-    })
-
-    it('should start at state 0', () => {
-      assert.strictEqual(new RandomWalk(0.5).state(), 0)
-    })
-  })
-
-  describe('path', () => {
-    it('should be integer-valued', () => {
-      const rw = new RandomWalk(0.5)
-      rw.seed(42)
-      const path = rw.path(200)
-      for (const x of path) {
-        assert.strictEqual(x, Math.floor(x))
-      }
-    })
-
-    it('should change by exactly +1 or -1 at each step', () => {
-      const rw = new RandomWalk(0.3)
-      rw.seed(7)
-      const path = rw.path(100)
-      for (let i = 1; i < path.length; i++) {
-        const diff = path[i] - path[i - 1]
-        assert(diff === 1 || diff === -1, `step ${i} diff = ${diff}`)
-      }
-    })
-  })
-
-  describe('.mean()', () => {
-    it('should return 0 for symmetric walk (p=0.5)', () => {
-      const rw = new RandomWalk(0.5)
-      assert.strictEqual(rw.mean(10), 0)
-    })
-
-    it('should return 0 at t = 0', () => {
-      const rw = new RandomWalk(0.7)
-      assert.strictEqual(rw.mean(0), 0)
-    })
-
-    it('should return NaN for t < 0', () => {
-      const rw = new RandomWalk(0.5)
-      assert(Number.isNaN(rw.mean(-1)))
-    })
-  })
-
-  describe('.variance()', () => {
-    it('should be reduced by bias (p != 0.5 has less variance than p = 0.5)', () => {
-      const sym = new RandomWalk(0.5)
-      const biased = new RandomWalk(0.3)
-      assert(biased.variance(10) < sym.variance(10))
-    })
-
-    it('should return 0 at t = 0', () => {
-      const rw = new RandomWalk(0.5)
-      assert.strictEqual(rw.variance(0), 0)
-    })
-
-    it('should return NaN for t < 0', () => {
-      const rw = new RandomWalk(0.5)
-      assert(Number.isNaN(rw.variance(-1)))
-    })
-  })
-
-  describe('.pdf()', () => {
-    it('should return NaN for t < 0', () => {
-      const rw = new RandomWalk(0.5)
-      assert(Number.isNaN(rw.pdf(0, -1)))
-    })
-
-    it('should return NaN for non-integer t', () => {
-      const rw = new RandomWalk(0.5)
-      assert(Number.isNaN(rw.pdf(0, 1.5)))
-    })
-
-    it('should return 0 for non-integer x', () => {
-      const rw = new RandomWalk(0.5)
-      assert.strictEqual(rw.pdf(0.5, 4), 0)
-    })
-
-    it('should return 0 when |x| > t', () => {
-      const rw = new RandomWalk(0.5)
-      assert.strictEqual(rw.pdf(3, 2), 0)
-    })
-
-    it('should return 0 when x and t have different parity', () => {
-      const rw = new RandomWalk(0.5)
-      // t=4 (even), x=1 (odd) — unreachable
-      assert.strictEqual(rw.pdf(1, 4), 0)
-    })
-
-    it('should return 1 at x=0, t=0 (initial point mass)', () => {
-      const rw = new RandomWalk(0.5)
-      assert.strictEqual(rw.pdf(0, 0), 1)
-    })
-
-    it('should return 0 at x=1, t=0', () => {
-      const rw = new RandomWalk(0.5)
-      assert.strictEqual(rw.pdf(1, 0), 0)
-    })
-
-    it('should sum to 1 over all reachable states at t=6', () => {
-      const rw = new RandomWalk(0.4)
-      let total = 0
-      for (let x = -6; x <= 6; x += 2) total += rw.pdf(x, 6)
-      // exact rational: sum of all binomial probabilities = 1
-      assert.closeTo(total, 1, 1e-10)
-    })
-  })
-
-  describe('.covariogram()', () => {
-    it('should be symmetric', () => {
-      const rw = new RandomWalk(0.7)
-      assert.closeTo(rw.covariogram(2, 4), rw.covariogram(4, 2), 1e-10)
-    })
-
-    it('should equal variance at s = t', () => {
-      const rw = new RandomWalk(0.6)
-      assert.closeTo(rw.covariogram(5, 5), rw.variance(5), 1e-10)
-    })
-
-    it('should return NaN for s < 0', () => {
-      const rw = new RandomWalk(0.5)
-      assert(Number.isNaN(rw.covariogram(-1, 2)))
-    })
-
-    it('should return NaN for t < 0', () => {
-      const rw = new RandomWalk(0.5)
-      assert(Number.isNaN(rw.covariogram(2, -1)))
-    })
-  })
-
-  describe('.reset()', () => {
-    it('should restore initial state to 0', () => {
-      const rw = new RandomWalk(0.5)
-      rw.seed(42)
-      for (let i = 0; i < 10; i++) rw.next()
-      rw.reset()
-      assert.strictEqual(rw.state(), 0)
-    })
-  })
-
-  describe('step distribution', () => {
-    it('should have mean and variance matching 2p-1 and 4p(1-p) across seeds', () => {
-      const p = 0.7
-      const n = 5000
-      // exact rational: a single ±1 step has mean = 2p-1, variance = 4p(1-p)
-      for (const seed of MOMENT_SEEDS) {
-        const rw = new RandomWalk(p)
-        rw.seed(seed)
-        assertSampleMoments(sampleSteps(rw, n), 2 * p - 1, 4 * p * (1 - p), seed)
-      }
-    })
-  })
-
-  describe('.marginal()', () => {
-    it('should have mean and variance matching mean(t) and variance(t)', () => {
-      const rw = new RandomWalk(0.6)
-      const marginal = rw.marginal(5)
-      assert.closeTo(marginal.mean(), rw.mean(5), 1e-10)
-      assert.closeTo(marginal.variance(), rw.variance(5), 1e-10)
-    })
-
-    it('should have pdf matching process.pdf(x, t) for on-parity x', () => {
-      const rw = new RandomWalk(0.6)
-      const marginal = rw.marginal(5)
-      assert.closeTo(marginal.pdf(3), rw.pdf(3, 5), 1e-10)
-      assert.closeTo(marginal.pdf(-1), rw.pdf(-1, 5), 1e-10)
-    })
-
-    it('should return 0 for off-parity x, matching process.pdf(x, t)', () => {
-      const rw = new RandomWalk(0.6)
-      const marginal = rw.marginal(4)
-      assert.strictEqual(marginal.pdf(1), rw.pdf(1, 4))
-      assert.strictEqual(marginal.pdf(1), 0)
-    })
-
-    it('should return 0 outside the support', () => {
-      const rw = new RandomWalk(0.6)
-      const marginal = rw.marginal(4)
-      assert.strictEqual(marginal.pdf(6), 0)
-      assert.strictEqual(marginal.pdf(-6), 0)
-    })
-
-    it('should invert cdf via quantile', () => {
-      const rw = new RandomWalk(0.6)
-      const marginal = rw.marginal(5)
-      assert.closeTo(marginal.q(marginal.cdf(1)), 1, 1e-10)
-      assert.closeTo(marginal.q(marginal.cdf(-3)), -3, 1e-10)
-    })
-
-    it('should sample within the support', () => {
-      const rw = new RandomWalk(0.6)
-      const marginal = rw.marginal(5)
-      marginal.seed(42)
-      for (const x of marginal.sample(50)) {
-        assert(x >= -5 && x <= 5, `x=${x} out of [-5, 5]`)
-        assert.strictEqual((5 + x) % 2, 0, `x=${x} has wrong parity`)
-      }
-    })
-
-    it('should sample values matching the marginal pmf shape (chi-square test)', () => {
-      // Bounds/parity checks above pass for any generator that merely stays inside
-      // [-n, n] with the right parity -- a flipped p (this.r.next() < this.p.p
-      // inverted) or a flipped step count (n - 2*heads instead of 2*heads - n) would
-      // still satisfy them undetected. A chi-square goodness-of-fit test against
-      // the ShiftedBinomial pmf catches a biased/inverted generator that those
-      // checks cannot.
-      const rw = new RandomWalk(0.6)
-      const marginal = rw.marginal(5)
-      marginal.seed(7)
-      const sample = marginal.sample(3000)
-      assert(chiTest(sample, x => marginal.pdf(x), 1))
-    })
-
-    it('should not throw at t = 0 and represent a point mass at 0', () => {
-      const rw = new RandomWalk(0.5)
-      const marginal = rw.marginal(0)
-      assert.strictEqual(marginal.pdf(0), 1)
-      assert.strictEqual(marginal.pdf(1), 0)
-    })
-
-    it('should throw for t < 0', () => {
-      const rw = new RandomWalk(0.5)
-      assert.throws(() => rw.marginal(-1), /non-negative integer/)
-    })
-
-    it('should throw for non-integer t', () => {
-      const rw = new RandomWalk(0.5)
-      assert.throws(() => rw.marginal(1.5), /non-negative integer/)
-    })
-
-    it('should support fitting the returned ShiftedBinomial instance to sampled data', () => {
-      const rw = new RandomWalk(0.7)
-      const marginal = rw.marginal(20)
-      marginal.seed(42)
-      const fitted = ShiftedBinomial.fit(marginal.sample(2000))
-      assert.instanceOf(fitted, ShiftedBinomial)
-      assert.strictEqual(fitted.params().n, 20)
-      assert.closeTo(fitted.params().p, 0.7, 0.05)
-    })
-
-    // ShiftedBinomial is private (decisions/0045) and only ever surfaces as the return value
-    // of RandomWalk.marginal(t), so these are its sole regression check for the six generic
-    // Distribution methods it inherits without overriding (hazard/survival/lnL/aic/bic/test) --
-    // a future change to the base class or to _cdf could silently break any of them otherwise.
-    it('should have survival(x) matching 1 - cdf(x)', () => {
-      const rw = new RandomWalk(0.6)
-      const marginal = rw.marginal(5)
-      assert.closeTo(marginal.survival(1), 1 - marginal.cdf(1), 1e-10)
-    })
-
-    it('should have hazard(x) matching pdf(x) / survival(x)', () => {
-      const rw = new RandomWalk(0.6)
-      const marginal = rw.marginal(5)
-      assert.closeTo(marginal.hazard(1), marginal.pdf(1) / marginal.survival(1), 1e-10)
-    })
-
-    it('should return a finite log-likelihood for a valid support sample', () => {
-      const rw = new RandomWalk(0.6)
-      const marginal = rw.marginal(5)
-      assert.isTrue(Number.isFinite(marginal.lnL([-3, -1, 1, 3, 5])))
-    })
-
-    it('should return finite aic/bic for a sampled data set', () => {
-      const rw = new RandomWalk(0.6)
-      const marginal = rw.marginal(5)
-      marginal.seed(1)
-      const sample = marginal.sample(50)
-      assert.isTrue(Number.isFinite(marginal.aic(sample)))
-      assert.isTrue(Number.isFinite(marginal.bic(sample)))
-    })
-
-    it('should return the discrete chi-square test result shape for a sampled data set', () => {
-      const rw = new RandomWalk(0.6)
-      const marginal = rw.marginal(5)
-      marginal.seed(7)
-      const result = marginal.test(marginal.sample(3000))
-      assert.isTrue(result.passed)
-      assert.isNumber(result.statistics)
-      assert.isNumber(result.pValue)
-    })
-  })
-
-  describe('.fit()', () => {
-    it('should recover p from a long simulated path across seeds', () => {
-      const p = 0.65
-      const n = 20000
-      // exact rational: p_hat is the sample mean of n iid Bernoulli(p) up-step indicators,
-      // so Var(p_hat) = p*(1-p)/n exactly (no asymptotic approximation needed).
-      const tolP = K_SIGMA * Math.sqrt(p * (1 - p) / n)
-      for (const seed of MOMENT_SEEDS) {
-        const rw = new RandomWalk(p)
-        rw.seed(seed)
-        const fitted = RandomWalk.fit(rw.path(n))
-        assert.instanceOf(fitted, RandomWalk)
-        assert.closeTo(fitted.params().p, p, tolP, `seed ${seed}`)
-      }
-    })
-
-    it('should throw when path has fewer than 2 states', () => {
-      assert.throws(() => RandomWalk.fit([0]), /at least 2 states/)
-    })
-
-    it('should throw when path is not an array', () => {
-      assert.throws(() => RandomWalk.fit(undefined), /at least 2 states/)
-    })
-
-    it('should throw when a step is not +1 or -1', () => {
-      assert.throws(() => RandomWalk.fit([0, 2, 1]), /not \+1 or -1/)
-    })
-
-    it('should throw when the estimated p is out of (0,1) (all up-steps)', () => {
-      assert.throws(() => RandomWalk.fit([0, 1, 2, 3]), /out of \(0,1\)/)
-    })
-
-    it('should throw for every 2-state path (a single Bernoulli trial can never yield a ' +
-      'fraction strictly between 0 and 1)', () => {
-      assert.throws(() => RandomWalk.fit([0, 1]), /out of \(0,1\)/)
-      assert.throws(() => RandomWalk.fit([0, -1]), /out of \(0,1\)/)
-    })
-
-    it('should succeed at 3 states (the minimum length where p_hat can land strictly inside (0,1))', () => {
-      const fitted = RandomWalk.fit([0, 1, 0])
-      assert.instanceOf(fitted, RandomWalk)
-      // exact rational: p_hat is the fraction of +1 steps, and [0, 1, 0] has increments
-      // [+1, -1], so p_hat = 1/2 → mean(t) = t*(2*0.5-1) = 0 for any t, and
-      // variance(1) = 4*0.5*(1-0.5)*1 = 1. Stays inline rather than moving to process-cases.js
-      // because fit() is a static factory, not an instance method the case file's shape can
-      // express.
-      assert.strictEqual(fitted.mean(4), 0)
-      assert.strictEqual(fitted.variance(1), 1)
-    })
-  })
-})
-
-// Externally-sourced reference values for the analytical methods, defined in ./process-cases.js.
-// Consumed by a plain inline loop rather than a shared runner module — see that file's header for
-// why a dist-runner.js-style abstraction is not proportionate at this scale (#1221).
-describe('process reference values', () => {
-  processCases.forEach(({ name, ctor: Ctor, refs }) => {
-    describe(name, () => {
-      refs.forEach(({ should, params, method, args, chain, chainArgs, expected, tol, source }) => {
-        // The method name is part of the title because `should` alone is not unique: Poisson's
-        // mean(t) and variance(t) are both "return lambda*t", and the per-method describe() blocks
-        // that used to disambiguate them are gone.
-        it(`${method}() should ${should}`, () => {
-          const proc = new Ctor(...params())
-          const value = proc[method](...args)
-          assert.closeTo(chain ? value[chain](...chainArgs) : value, expected, tol, source)
-        })
-      })
-    })
-  })
-})
